% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ags.R
\name{ags}
\alias{ags}
\alias{ags_link}
\alias{is_ags}
\alias{ags_parse}
\title{AGS geolinker}
\usage{
ags_link(.data, id_col, level, country = "DEU", ...)

is_ags(ids)

ags_parse(ids)
}
\arguments{
\item{.data}{A dataframe with at least one column named according to the
\code{id_col} argument.}

\item{id_col}{Column in \code{.data} holding the spatial identifiers
in question.}

\item{level}{Vector of geographic levels. Can be \code{"sta"}, \code{"lan"},
\code{"rbz"}, \code{"krs"}, \code{"vwg"} or \code{"gem"}. See
\code{\link[ffm]{bkg_admin}} for details.}

\item{country}{Character vector of unique ISO-3 country codes contained
by \code{ids}. Always \code{"DEU"}.}

\item{...}{Further arguments passed \code{\link[ffm]{bkg_admin}}.
Particularly attribute filters or spatial filters.}

\item{ids}{A character vector of IDs.}
}
\value{
\code{is_ags}: A logical vector of length \code{length(ids)}.

\code{ags_parse}: A list with fields \code{country} and \code{level}.

\code{ags_link}: An sf dataframe containing \code{.data} and an added
geometry column.
}
\description{
Default geolinker registered through \code{\link{register_geolinker}}.
These functions are not intended to be used directly but through
\code{\link{enrich}}!

Contains functions to link, parse and validate geometries from the
German official municipality key (AGS,
\url{https://en.wikipedia.org/wiki/Community_Identification_Number}).

Requires the \href{https://jslth.github.io/ffm/}{\code{ffm}}
package to be installed.
}
\examples{
\dontshow{if (requireNamespace("ffm")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{# Link Hamburg
enrich("02", linker = "ags")}
\dontshow{\}) # examplesIf}
}
