% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertile_processor.R
\name{as.data.table.raster}
\alias{as.data.table.raster}
\title{Convert a raster to data.table object}
\usage{
\method{as.data.table}{raster}(
  x,
  row.names = NULL,
  optional = FALSE,
  xy = FALSE,
  inmem = raster::canProcessInMemory(x, 2),
  ...
)
}
\arguments{
\item{x}{an object of class `raster`.}
}
\description{
This function acts just like the `raster::as.data.frame()` object
}
