% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonalstats_plugin.R
\name{compute_zonalstats}
\alias{compute_zonalstats}
\title{A function to compute zonal statistics from multiple rasters into one shapefile
object}
\usage{
compute_zonalstats(shp_dt, raster_objs, extract_fun, name_set, weight_raster)
}
\arguments{
\item{shp_dt}{an sf/dataframe with polygons}

\item{raster_objs}{a list of raster objects to be extracted from}

\item{extract_fun}{function to be extracted}

\item{name_set}{a nameset which will be created for the variables}

\item{weight_raster}{a raster object or a list of weighting rasters or the
equal in length to \code{raster_objs}}
}
\description{
A function to compute zonal statistics from multiple rasters into one shapefile
object
}
\details{
A parallelization option will be added at
some point
}
