% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_ntl_internals.R
\name{construct_month_link}
\alias{construct_month_link}
\title{Construct URL to the month night time lights data}
\usage{
construct_month_link(
  year,
  month,
  version,
  slc_type,
  no_tile,
  link_base = "https://eogdata.mines.edu/nighttime_light"
)
}
\arguments{
\item{year}{an integer, the year}

\item{month}{an integer, the month}

\item{version}{a character, the NTL version "v10"}

\item{slc_type}{a character, "vcmslcfg"}

\item{no_tile}{a logical, always set to FALSE}

\item{link_base}{a character, the base link to the NTL EOG database}
}
\description{
Construct URL to the month night time lights data
}
