% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_ntl_internals.R
\name{construct_year_link}
\alias{construct_year_link}
\title{Construct URL to the annual night time lights data}
\usage{
construct_year_link(
  year,
  version,
  link_base = "https://eogdata.mines.edu/nighttime_light/annual/"
)
}
\arguments{
\item{year}{an integer, the year}

\item{version}{a character, the NTL version ("v10", "v20", "v21" and "v22")}

\item{link_base}{a character, the base link to the NTL EOG database}
}
\description{
Construct URL to the annual night time lights data
}
