% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_landcover.R
\name{create_class_mapping}
\alias{create_class_mapping}
\title{Create class mapping from STAC features}
\usage{
create_class_mapping(it_obj)
}
\arguments{
\item{it_obj}{STAC item object}
}
\value{
Class mapping dataframe
}
\description{
Create class mapping from STAC features
}
