% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_landcover.R
\name{create_default_results}
\alias{create_default_results}
\title{Helper functions
Create default results dataframe for a year}
\usage{
create_default_results(sf_obj, year)
}
\arguments{
\item{sf_obj}{Spatial object}

\item{year}{Year for which to create default results}
}
\value{
A dataframe with default values
}
\description{
Helper functions
Create default results dataframe for a year
}
