% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoareacreator_plugin.R
\name{create_query_bbox}
\alias{create_query_bbox}
\title{A function to create an osmdata package readable bounding box (bbox)
with a buffer distance}
\usage{
create_query_bbox(
  shp_dt = NULL,
  area_name,
  buffer_dist = c(0, 0, 0, 0),
  metric_crs = FALSE,
  osm_crs = 4326
)
}
\description{
A function to create an osmdata package readable bounding box (bbox)
with a buffer distance
}
