% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_generalhelpers.R
\name{download_dictionary}
\alias{download_dictionary}
\title{A function to create dictionary of file extensions and the functions that
read them so that a do.call can be applied to read in the results}
\usage{
download_dictionary()
}
\description{
A function to create dictionary of file extensions and the functions that
read them so that a do.call can be applied to read in the results
}
