% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_landcover.R
\name{download_raster}
\alias{download_raster}
\title{Download a raster from URL}
\usage{
download_raster(url, output_path)
}
\arguments{
\item{url}{URL to download from}

\item{output_path}{Path to save the downloaded file}
}
\value{
Path to downloaded file or NULL if download failed
}
\description{
Download a raster from URL
}
