% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_generalhelpers.R
\name{download_reader}
\alias{download_reader}
\title{A function to download and read in a file from the internet}
\usage{
download_reader(url, ...)
}
\arguments{
\item{url}{a link URL}

\item{...}{additional arguments to be used in `download.file()`}
}
\description{
A function to download and read in a file from the internet
}
