% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_generalhelpers.R
\name{download_worker}
\alias{download_worker}
\title{Basic function for downloading}
\usage{
download_worker(dsn, url)
}
\arguments{
\item{dsn}{a folder location}

\item{url}{a link URL}
}
\description{
Basic function for downloading
}
