% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrich}
\alias{enrich}
\title{Enrich data with geometries}
\usage{
enrich(
  .data,
  id_col = "id",
  linker = NULL,
  country = NULL,
  level = NULL,
  iso3_scheme = NULL,
  iso3_auto = TRUE,
  iso3_default = "naturalearth",
  crs = 3035,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{.data}{A dataframe with at least one column named according to the
\code{id_col} argument.}

\item{id_col}{Column in \code{.data} holding the spatial identifiers
in question.}

\item{linker}{The type of spatial identifier present in
\code{.data[[id_col]]}. Can be either of:
\itemize{
\item{\code{NULL}, in which case the geolinker is guessed based on the
ID format.}
\item{The name of a geolinker registered by
\code{\link{register_geolinker}}. By default, this includes at least
\code{"gadm"}, \code{"unhcr"}, \code{"nuts"}, \code{"inspire"}, \code{"lau"},
\code{"ags"}, \code{"naturalearth"}, \code{"fips"}, and \code{"postcode"}.
See the "See also" section for their documentation.}
\item{A function taking the arguments \code{.data} and \code{id_col}.
The output should be an \code{sf} dataframe with linked data.
This is equivalent of registering a link function using
\code{\link{register_geolinker}} and then passing its name to this
argument.}
}}

\item{country}{For linkers where a country is required (e.g., \code{"gadm"}
or \code{"nuts"}), specifies a character vector of countries. If
\code{NULL}, the countries are parsed from the ID strings. Note that
some geolinkers might not have a parse function but require country
specifications. In this case, this argument may be necessary.}

\item{level}{For linkers where a geographic level is required (e.g.,
\code{"gadm"} or \code{"nuts"}), specifies a character vector of levels.
If \code{NULL}, the levels are parsed from the ID strings. Note that
some geolinkers might not have a parse function but require country
specifications. In this case, this argument may be necessary.}

\item{iso3_scheme}{If \code{.data[[id_col]]} contains any country code that
can be parsed by \code{\link[countrycode]{countrycode}}, specifies the
country code scheme that the IDs follow, e.g. \code{"eurostat"} or
\code{"un"}. See \code{\link[countrycode]{codelist}} for possible code
schemes. \code{enrich} will convert them to ISO-3 and merge them using the
geometry source specified in \code{iso3_default}. If \code{NULL} (default)
and \code{iso3_auto} is \code{TRUE}, tries to guess the code scheme.
Otherwise, leaves the IDs as-is.}

\item{iso3_auto}{If \code{TRUE} (default), tries to automatically convert
identifiers to ISO-3 using \code{\link[countrycode]{guess_field}}. This
is necessary if \code{id_col} holds any country codes specified in
\code{\link[countrycode]{codelist}} (e.g., GAUL, FAO, M49, FIPS, Eurostat).}

\item{iso3_default}{In case \code{id_col} holds country codes that are
converted to ISO-3, specifies which geometry source to use. Defaults to
\code{"naturalearth"}, but can also take \code{"gadm"} and
\code{"unhcr"}. \code{"gadm"}, \code{"unhcr"}, and \code{"geoboundaries"}
are generally much slower than \code{"naturalearth"} but may provide
more appropriate geometries for certain use cases. \code{"unhcr"} may
fail with many unique country codes.}

\item{crs}{Output coordinate reference system that merged data are
transformed to. Defaults to EPSG:3035 (ETRS89 / LAEA Europe).}

\item{verbose}{If \code{TRUE}, prints informative messages about the
guessing process.}

\item{...}{Further arguments passed to the linking function.}
}
\value{
An sf dataframe (or tibble if possible) with the original data
of \code{.data} and a merged geometry column.
}
\description{
Given a dataframe with spatial identifiers NUTS codes or ISO-3
country codes links them with their spatial representation. If you have a
dataframe with an identifier and you don't know which type of identifier it
is, this function can also guess it for you.
}
\seealso{
Individual geolinkers:
\code{\link{naturalearth}},
\code{\link{gadm}},
\code{\link{unhcr}},
\code{\link{geoboundaries}},
\code{\link{nuts}},
\code{\link{inspire}},
\code{\link{lau}},
\code{\link{ags}},
\code{\link{fips}},
\code{\link{postcode}}
}
