% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_landcover.R
\name{ensure_crs_4326}
\alias{ensure_crs_4326}
\title{Ensure coordinates are in CRS 4326}
\usage{
ensure_crs_4326(data)
}
\arguments{
\item{data}{Spatial data object}
}
\value{
Spatial data object with CRS 4326
}
\description{
Ensure coordinates are in CRS 4326
}
