% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_geolinkers}
\alias{export_geolinkers}
\alias{import_geolinkers}
\title{Export and import geolinkers}
\usage{
export_geolinkers(path, name = NULL, ...)

import_geolinkers(file)
}
\arguments{
\item{path}{A directory path where to store the geolinkers.}

\item{name}{The name of the exported file. If \code{NULL}, generates
a random name.}

\item{...}{Further arguments passed to \code{\link{saveRDS}}.}

\item{file}{A file path to import. Must point to a file previously created
by \code{export_geolinkers}.}
}
\value{
\code{export_geolinkers} returns the output file path invisibly.
\code{import_geolinkers} returns \code{NULL} invisibly.
}
\description{
\code{export_geolinkers} exports a database of geolinkers registered
using \code{register_geolinkers} to a file.

\code{import_geolinkers} imports an exported database of geolinkers to
the current R session.

\strong{Note:} To prevent unexpected behavior, \code{import_geolinkers}
\emph{always} overwrites all existing geolinkers in a session. In
interactive sessions, the function will warn about this behavior.
}
\examples{
# in session A:
geodir <- tempdir()
export_geolinkers(geodir, name = "example")

# in session B:
import_geolinkers(file.path(geodir, "example"))
}
