% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_landcover.R
\name{extract_and_process_values}
\alias{extract_and_process_values}
\title{Extract and process raster values for polygons}
\usage{
extract_and_process_values(rast_obj, sf_obj, class_mapping, results)
}
\arguments{
\item{rast_obj}{Raster object}

\item{sf_obj}{Spatial object}

\item{class_mapping}{Class mapping dataframe}

\item{results}{Results dataframe to populate}
}
\value{
Updated results dataframe
}
\description{
Extract and process raster values for polygons
}
