% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_landcover.R
\name{filter_features}
\alias{filter_features}
\title{Filter STAC features based on date and bounding box}
\usage{
filter_features(feature, start_date, end_date)
}
\arguments{
\item{feature}{STAC feature}

\item{start_date}{Start date}

\item{end_date}{End date}
}
\value{
Boolean indicating if feature should be included
}
\description{
Filter STAC features based on date and bounding box
}
