% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\name{fips}
\alias{fips}
\alias{fips_link}
\title{FIPS geolinker}
\usage{
fips_link(.data, id_col, ...)
}
\arguments{
\item{.data}{A dataframe with at least one column named according to the
\code{id_col} argument.}

\item{id_col}{Column in \code{.data} holding the spatial identifiers
in question.}

\item{...}{Not used.}
}
\value{
An sf dataframe containing \code{.data} and an added geometry column.
}
\description{
Default geolinker registered through \code{\link{register_geolinker}}.
These functions are not intended to be used directly but through
\code{\link{enrich}}!

Contains functions to link FIPS geometries from the US census. Provides no
functions for validating and parsing as FIPS codes are usually pretty
generic numbers.

Requires the \href{https://fipio.justinsingh.me/}{\code{fipio}}
package to be installed.
}
\examples{
\dontshow{if (requireNamespace("fipio")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{# Link Alaska
enrich("02", linker = "fips")}
\dontshow{\}) # examplesIf}
}
