% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadm.R
\name{gadm}
\alias{gadm}
\alias{gadm_link}
\alias{gadm_parse}
\alias{is_gadm}
\title{GADM geolinker}
\usage{
gadm_link(.data, id_col, country, level = 0, resolution = 1, ...)

gadm_parse(ids)

is_gadm(ids)
}
\arguments{
\item{.data}{A dataframe with at least one column named according to the
\code{id_col} argument.}

\item{id_col}{Column in \code{.data} holding the spatial identifiers
in question.}

\item{country}{Character vector of unique ISO-3 country codes contained
by \code{ids}. Only geometries within these countries are returned.}

\item{level}{Vector of geographic levels. Only geometries that adhere to
these levels are returned. Can be an integer between 0 (country level)
and 4 (municipality level).}

\item{...}{Further arguments passed to \code{\link[geodata:gadm]{geodata::gadm}}.}

\item{ids}{A character vector of IDs.}
}
\value{
\code{is_gadm}: A logical vector of length \code{length(ids)}.

\code{gadm_parse}: A list with fields \code{country} and \code{level}.

\code{gadm_link}: An sf dataframe containing \code{.data} and an added
geometry column.
}
\description{
Default geolinker registered through \code{\link{register_geolinker}}.
These functions are not intended to be used directly but through
\code{\link{enrich}}!

Contains functions to link, parse and validate geometries from the Database
of Global Administrative Areas (GADM, \url{https://gadm.org/}).
}
\examples{
\donttest{enrich("MCO", linker = "gadm")}
}
