% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoareacreator_plugin.R
\name{gengrid2}
\alias{gengrid2}
\title{Another function to create a gridified shapefile and extract a raster if specified}
\usage{
gengrid2(shp_dt, grid_size, sqr = TRUE)
}
\arguments{
\item{shp_dt}{an object of class 'sf' or 'sfc'}

\item{grid_size}{numeric of length 1; representing the desired size of the grid in meters}

\item{sqr}{logical; if TRUE, a square grid is created. If FALSE, a hexagonal polygon grid is created}
}
\description{
This function takes in only a shapefile and creates a square or hexagon polygon grid based on a specified
grid size
}
