% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoboundaries.R
\name{geoboundaries}
\alias{geoboundaries}
\alias{gb_link}
\title{geoBoundaries geolinker}
\usage{
gb_link(.data, id_col, country, ...)
}
\arguments{
\item{.data}{A dataframe with at least one column named according to the
\code{id_col} argument.}

\item{id_col}{Column in \code{.data} holding the spatial identifiers
in question.}

\item{country}{Character vector of unique ISO-3 country codes contained
by \code{ids}. Only geometries within these countries are returned.}

\item{...}{Further arguments passed to
\code{\link[rgeoboundaries]{geoboundaries}}.}
}
\value{
An sf dataframe containing \code{.data} and an added geometry column.
}
\description{
Default geolinker registered through \code{\link{register_geolinker}}.
These functions are not intended to be used directly but through
\code{\link{enrich}}!

Contains functions to link geometries from geoBoundaries
(\url{https://www.geoboundaries.org/}). Provides no functions for validating
and parsing as geoBoundaries does not provide its own admin IDs.

Requires the \href{https://github.com/wmgeolab/rgeoboundaries}{rgeoboundaries}
package to be installed.
}
\examples{
\dontshow{if (getFromNamespace("run_examples", ns = "geolink")()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Link Monaco
enrich("MCO", linker = "geoboundaries")
\dontshow{\}) # examplesIf}
}
