% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_CMIP6}
\alias{geolink_CMIP6}
\title{Download CMIP6 climate model data}
\usage{
geolink_CMIP6(
  start_date,
  end_date,
  scenario,
  desired_models,
  shp_dt,
  shp_fn = NULL,
  grid_size = NULL,
  survey_dt = NULL,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  buffer_size = NULL,
  extract_fun = "mean",
  survey_crs = 4326
)
}
\arguments{
\item{start_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{end_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{scenario}{A scenario has to be selected and can be one of "historical", "ssp245" or "ssp585"}

\item{desired_models}{The name or names in a list, of the desired model(s) required for the analysis, for example
("ACCESS-CM-2" or ["ACCESS-CM-2","UKESM1-0-LL" ]). See the cmip6:model summary in the STAC collection for a full list of models.}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons representing the study area.}

\item{shp_fn}{A character, file path for the shapefile (.shp) to be read (for STATA users only).}

\item{grid_size}{A numeric, the grid size to be used in meters for analyzing the climate model data.}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey with latitude and longitude values (optional).}

\item{survey_fn}{A character, file path for geocoded survey (.dta format) (for STATA users only & if use_survey is TRUE) (optional).}

\item{survey_lat}{A character, latitude variable from survey (for STATA users only & if use_survey is TRUE) (optional).}

\item{survey_lon}{A character, longitude variable from survey (for STATA users only & if use survey is TRUE) (optional).}

\item{buffer_size}{A numeric, the buffer size to be used around each point in the survey data, in meters (optional).}

\item{extract_fun}{A character, a function to be applied in extraction of raster into the shapefile.
Default is "mean". Other options are "sum", "min", "max", "sd", "skew" and "rms" (optional).}

\item{survey_crs}{An integer, the Coordinate Reference System (CRS) for the survey data. Default is 4326 (WGS84) (optional).}
}
\value{
A processed data frame based on the input parameters and downloaded data.
}
\description{
This function downloads and processes CMIP6 (Coupled Model Intercomparison Project Phase 6)
climate model data for a specific scenario and desired model.
It allows for further analysis of the data in conjunction with geographic data.
}
\examples{
\donttest{

# Example usage
df <- geolink_CMIP6(start_date = "2019-01-01", end_date = "2019-12-31",
                      scenario = "ssp245", desired_models = "UKESM1-0-LL",
                       shp_dt = shp_dt[shp_dt$ADM1_EN == "Abia",])


}

}
