% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_electaccess}
\alias{geolink_electaccess}
\title{Download high resolution electrification access data from HREA}
\usage{
geolink_electaccess(
  start_date = NULL,
  end_date = NULL,
  shp_dt = NULL,
  shp_fn = NULL,
  grid_size = 1000,
  survey_dt,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  buffer_size = NULL,
  extract_fun = "mean",
  survey_crs = 4326
)
}
\arguments{
\item{start_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{end_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons}

\item{shp_fn}{A character, file path for the shapefile (.shp) to be read (for STATA users only)}

\item{grid_size}{A numeric, the grid size to be used in meters}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey i.e. a household survey with latitude and longitude values.}

\item{survey_fn}{A character, file path for geocoded survey (.dta format) (for STATA users only & if use_survey is TRUE)}

\item{survey_lat}{A character, latitude variable from survey (for STATA users only & if use_survey is TRUE)}

\item{survey_lon}{A character, longitude variable from survey (for STATA users only & if use survey is TRUE)}

\item{extract_fun}{A character, a function to be applied in extraction of raster into the shapefile.
Default is mean. Other options are "sum", "min", "max", "sd", "skew" and "rms".}
}
\description{
Download high resolution electrification access data from HREA
}
\details{
Details for the dataset can be found here: https://hrea.isr.umich.edu/
}
\examples{
\donttest{


df <- geolink_electaccess(shp_dt = shp_dt[shp_dt$ADM1_EN == "Abia",])

}

}
