% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_get_poi}
\alias{geolink_get_poi}
\title{Download points of interest from OSM data using open street maps API.}
\usage{
geolink_get_poi(
  osm_feature_category,
  osm_feature_subcategory,
  shp_dt,
  shp_dsn = NULL,
  buffer = NULL,
  stata = FALSE
)
}
\arguments{
\item{osm_feature_category}{A character, refering to the osm key wiki page, please see details below}

\item{osm_feature_subcategory}{A character, refering to the osm key wiki page, please see details below}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons}

\item{shp_dsn}{A link to location of shapefile for stat users}

\item{buffer}{buffer area around shapefile}

\item{stata}{A flag for stata users}
}
\description{
Download points of interest from OSM data using open street maps API.
}
\details{
Details for feature category and sub-category can be found here: https://wiki.openstreetmap.org/wiki/Map_features
}
\examples{
\donttest{



df <- geolink_get_poi(osm_feature_category = "building",
                      osm_feature_subcategory = "farm",
                      shp_dt = shp_dt)

}
}
