% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_ntl}
\alias{geolink_ntl}
\title{Download and Merge Monthly and Annual Night Time Light data into geocoded surveys using BlackMarbleR}
\usage{
geolink_ntl(
  time_unit = "annual",
  start_date,
  end_date,
  username,
  password,
  variable = NULL,
  quality_flag_rm = NULL,
  shp_dt = NULL,
  shp_fn = NULL,
  grid_size = NULL,
  survey_dt = NULL,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  extract_fun = "mean",
  buffer_size = NULL,
  survey_crs = 4326,
  return_raster = FALSE,
  weight_raster = NULL,
  add_n_pixels = TRUE,
  check_all_tiles_exist = TRUE,
  download_method = "httr",
  h5_dir = NULL
)
}
\arguments{
\item{time_unit}{A character, either "month" or "annual" monthly or annual nighttime lights aggregates
are to be estimated}

\item{start_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{end_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{username}{A character, NASA Earthdata username for automatic bearer token generation}

\item{password}{A character, NASA Earthdata password for automatic bearer token generation}

\item{variable}{A character, the Black Marble variable to extract. If NULL, uses default variables:
"Gap_Filled_DNB_BRDF-Corrected_NTL" for daily, "NearNadir_Composite_Snow_Free" for monthly/annual}

\item{quality_flag_rm}{Quality flag values to set to NA (default: NULL). See BlackMarbleR documentation.}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons}

\item{shp_fn}{A character, file path for the shapefile (.shp) to be read (for STATA users only)}

\item{grid_size}{A numeric, the grid size to be used in meters}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey i.e.
a household survey with latitude and longitude values.}

\item{survey_fn}{A character, file path for geocoded survey (.dta format) (for STATA users only)}

\item{survey_lat}{A character, latitude variable from survey (for STATA users only &
if use_survey is TRUE)}

\item{survey_lon}{A character, longitude variable from survey (for STATA users only &
if use survey is TRUE)}

\item{extract_fun}{A character, a function to be applied in extraction of raster into the shapefile.}

\item{buffer_size}{A numeric, the size of the buffer for \code{survey_dt} or \code{survey_fn} in meters.}

\item{survey_crs}{A numeric, the default is 4326
Default is mean. Other options are "sum", "min", "max", "sd" and others supported by exactextractr.}

\item{return_raster}{logical, default is FALSE, if TRUE a raster will be returned ONLY. The resulting
raster is cropped to the extent of \code{shp_dt} if \code{shp_dt} is specified.}

\item{weight_raster}{a raster object of class \code{spatRaster} or a list of \code{spatRaster} objects}

\item{add_n_pixels}{logical, whether to add pixel count information when extracting (default: TRUE)}

\item{check_all_tiles_exist}{logical, check if all tiles exist for the region (default: TRUE)}

\item{download_method}{character, method for downloading: "httr" or "wget" (default: "httr")}

\item{h5_dir}{A character, directory path where H5 files should be stored. If NULL, uses temporary directory.}
}
\description{
Download night lights luminosity at monthly/annual intervals for a specified period
The data is downloaded using BlackMarbleR package and combined with shapefile and/or survey data provided
by the user
}
\examples{

\dontrun{
\donttest{



df <- geolink_ntl(
  time_unit = "annual",
  start_date = "2020-01-01",
  end_date = "2020-01-10",
  username = username,
  password = password,
  shp_dt = shp_dt[shp_dt$ADM1_EN == "Abia",],
  extract_fun = "mean")

}}

}
