% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_population}
\alias{geolink_population}
\title{Download and Merge Annual Population data into geocoded surveys}
\usage{
geolink_population(
  start_year = NULL,
  end_year = NULL,
  iso_code,
  UN_adjst = NULL,
  constrained = NULL,
  bespoke = NULL,
  version = NULL,
  shp_dt = NULL,
  shp_fn = NULL,
  grid_size = NULL,
  survey_dt = NULL,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  buffer_size = NULL,
  extract_fun = "mean",
  survey_crs = 4326,
  file_location = tempdir()
)
}
\arguments{
\item{start_year}{A numeric specifying the start year}

\item{end_year}{A numeric specifying the end year, if only one year is required then enter start year into}

\item{iso_code}{A character, specifying the iso code for country to download data from}

\item{UN_adjst}{A character 'Y' if want 2020 UN adjusted dataset returned, if not leave blank.}

\item{constrained}{A character 'Y' if want constrained dataset returned, if not leave blank.}

\item{bespoke}{A character 'Y' if want bespoke dataset returned, if not leave blank. Version also should be specified.}

\item{version}{A character such as v2.0 or v2.1 which correlates to the iso_code provided for bespoke datasets.}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons}

\item{shp_fn}{A character, file path for the shapefile (.shp) to be read (for STATA users only)}

\item{grid_size}{A numeric, the grid size to be used in meters}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey i.e.
a household survey with latitude and longitude values.}

\item{survey_fn}{A character, file path for geocoded survey (.dta format) (for STATA users only &
if use_survey is TRUE)}

\item{survey_lat}{A character, latitude variable from survey (for STATA users only &
if use_survey is TRUE)}

\item{survey_lon}{A character, longitude variable from survey (for STATA users only &
if use survey is TRUE)}

\item{extract_fun}{A character, a function to be applied in extraction of raster into the shapefile.
Default is mean. Other options are "sum", "min", "max", "sd", "skew" and "rms".}

\item{survey_crs}{A numeric, the default is 4326}

\item{file_location}{A path to the folder where the downloaded data should be stored.}

\item{buffer_survey}{A logical, specify TRUE if interested in estimating a statistic based on distance
from the survey location.}
}
\description{
Download Population data from the World Pop dataset at annual intervals for a specified period
The data is downloaded in raster format and combined with shapefile and/or survey data provided
by the user. Source data: https://www.worldpop.org/
}
\details{
The individual country datasets use modelling methods found it Stevens et al. The 'Global per country 2000-2020'
The datasets provided are the outputs of a project utilising various modelling methods to develop consistent 100m
resolution population count datasets for all countries of the World for each year 2000-2020.

Population data is sourced from WorldPop.
The data is extracted into a shapefile provided by user. An added service for tesselating/gridding
the shapefile is also provided for users that need this data further analytics that require
equal area zonal statistics. Shapefile estimates at the grid or native polygon level is a
permitted final output. However, a geocoded survey with population estimates are the end goal
if the user also chooses. The function will merge shapefile polygons (either gridded or
native polygons) with the location of the survey units i.e. population estimates for the
locations of the units within the survey will be returned. The function is also set up for
stata users and allows the user to pass file paths for the household survey `survey_fn`
(with the lat and long variable names `survey_lon` and `survey_lat`) as well. This is requires
a .dta file which is read in with `haven::read_dta()` package. Likewise, the user is permitted
to pass a filepath for the location of the shapefile `shp_fn` which is read in with the
`sf::read_sf()` function.

'@import rstac reticulate terra raster osmdata sf  geodata rvest httr
}
\examples{
\donttest{



df <- geolink_population(start_year = 2018,
                           end_year = 2019,
                          iso_code = "NGA",
                           UN_adjst = "N",
                           constrained = "N",
                           shp_dt = shp_dt[shp_dt$ADM1_EN == "Abia",],
                           grid_size = 1000,
                           extract_fun = "mean",
                           file_location = "/Users/nikos/Documents/temp/nga_pop")

}
}
