% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_vegindex}
\alias{geolink_vegindex}
\title{Download vegetation index data}
\usage{
geolink_vegindex(
  start_date,
  end_date,
  indicator = "NDVI",
  shp_dt = NULL,
  shp_fn = NULL,
  resolution = NULL,
  grid_size = NULL,
  survey_dt = NULL,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  buffer_size = NULL,
  extract_fun = "mean",
  survey_crs = 4326,
  return_raster = FALSE,
  weight_raster = NULL
)
}
\arguments{
\item{start_date}{A character object with the start date; it must be specified as "yyyy-mm-dd"}

\item{end_date}{A character object with the end date; it must be specified as "yyyy-mm-dd"}

\item{indicator}{A character object with the vegetation index to be used; it must be either "NDVI" or "EVI".}

\item{shp_dt}{An object of class 'sf', which contains polygons or multipolygons representing the study area.}

\item{shp_fn}{A character, file path for the shapefile (.shp) to be read (for STATA users only).}

\item{resolution}{A numeric, the resolution to be used in meters for extracting the vegetation index data.}

\item{grid_size}{A numeric, the grid size to be used to extract the data.}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey with latitude and longitude values (optional).}

\item{survey_fn}{A character, file path for geocoded survey (.dta format) (for STATA users only & if use_survey is TRUE) (optional).}

\item{survey_lat}{A character, latitude variable from survey (for STATA users only & if use_survey is TRUE) (optional).}

\item{survey_lon}{A character, longitude variable from survey (for STATA users only & if use survey is TRUE) (optional).}

\item{buffer_size}{A numeric, the buffer size to be used around each point in the survey data, in meters (optional).}

\item{extract_fun}{A character, a function to be applied in extraction of raster into the shapefile.
Default is "mean". Other options are "sum", "min", "max", "sd", "skew" and "rms" (optional).}

\item{survey_crs}{An integer, the Coordinate Reference System (CRS) for the survey data. Default is 4326 (WGS84) (optional).}

\item{return_raster}{logical, default is FALSE, if TRUE a raster will be returned ONLY. The resulting
raster is cropped to the extent of \code{shp_dt} if \code{shp_dt} is specified. Otherwise, full raster from
source is downloaded.}

\item{weight_raster}{a raster object of class \code{spatRaster} or a list of \code{spatRaster} objects}
}
\value{
A processed data frame based on the input parameters and downloaded data.
}
\description{
This function download vegetation index data, either NDVI or EVI, based on start and end dates.
}
\examples{
\dontrun{
\donttest{

 #example usage
df <- geolink_vegindex(shp_dt = shp_dt[shp_dt$ADM1_EN ==  "Abia",],
                             start_date = "2019-01-01",
                             end_date = "2019-12-31",
                             indicator = "NDVI",
                             extract_fun = "mean",
                             buffer_size = 1000,
                             survey_crs = 4326)


}}
}
