% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_worldclim}
\alias{geolink_worldclim}
\title{Download WorldClim climate data}
\usage{
geolink_worldclim(
  iso_code,
  var,
  res,
  shp_dt,
  shp_fn = NULL,
  grid_size = 1000,
  survey_dt,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  buffer_size = NULL,
  extract_fun = "mean",
  survey_crs = 4326
)
}
\arguments{
\item{iso_code}{A character, specifying the iso code for country to download data from}

\item{var}{A character, the variable of interest (e.g., "temperature", "precipitation").}

\item{res}{A character, the resolution of the data (e.g., "2.5m", "5m").}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons representing the study area.}

\item{shp_fn}{A character, file path for the shapefile (.shp) to be read (for STATA users only).}

\item{grid_size}{A numeric, the grid size to be used in meters for analyzing the climate data.}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey with latitude and longitude values (optional).}

\item{survey_fn}{A character, file path for geocoded survey (.dta format) (for STATA users only & if use_survey is TRUE) (optional).}

\item{survey_lat}{A character, latitude variable from survey (for STATA users only & if use_survey is TRUE) (optional).}

\item{survey_lon}{A character, longitude variable from survey (for STATA users only & if use survey is TRUE) (optional).}

\item{buffer_size}{A numeric, the buffer size to be used around each point in the survey data, in meters (optional).}

\item{extract_fun}{A character, a function to be applied in extraction of raster into the shapefile.
Default is "mean". Other options are "sum", "min", "max", "sd", "skew" and "rms" (optional).}

\item{survey_crs}{An integer, the Coordinate Reference System (CRS) for the survey data. Default is 4326 (WGS84) (optional).}
}
\value{
A processed data frame or object based on the input parameters and downloaded data.
}
\description{
This function downloads WorldClim climate data for a specific variable and resolution. It allows for further analysis of climate patterns in a given area.
}
\examples{
\donttest{

# Example usage
df <-geolink_worldclim(iso_code ="NGA",
                       var = 'tmax',
                       res= 2.5,
                       shp_dt = shp_dt[shp_dt$ADM1_EN == "Abia",],
                       grid_size = 1000,
                       extract_fun = "mean")
}
}
