% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_chirps.R
\name{get_annual_chirps}
\alias{get_annual_chirps}
\title{Download annual rainfall chirp data}
\usage{
get_annual_chirps(
  start_year,
  end_year,
  link_base = "https://data.chc.ucsb.edu/products/CHIRPS-2.0/",
  cores = 1L
)
}
\arguments{
\item{start_year}{An integer for the first year CHIRPS data to download}

\item{end_year}{An integer for the final year in the series of annual rasters to download}

\item{link_base}{the link to the daily chirps data (no need to change this)}

\item{cores}{the number of PC cores to employ in pulling the data in parallel}

\item{resolution}{A character, gridded resolution of the data (options area "p05" i.e.
0.05 degree resolution and p25 for 0.25)}

\item{dsn}{data source name (a local folder where the data is to be stored)}
}
\description{
Pull rainfall data from the CHIRPS data at monthly intervals for a specified period
The data is downloaded in raster format (.tif)
}
