% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_ntl.R
\name{get_month_ntl}
\alias{get_month_ntl}
\title{Download Monthly Night Time Lights Data}
\usage{
get_month_ntl(
  start_date,
  end_date,
  version,
  no_tile,
  slc_type,
  indicator = c("avg_rade9h", "cf_cvg", "cvg", "avg_rade9h.masked"),
  link_base = "https://eogdata.mines.edu/nighttime_light"
)
}
\arguments{
\item{start_date}{A character, the starting date of interest}

\item{end_date}{A character, the ending date of interest}

\item{version}{A string character, for now only "v10" is supported more to come later}

\item{no_tile}{character, tile or not (see usage within the upstream NTL function)}

\item{slc_type}{A character, adjustment type made to the NTL data either "vcmcfg" or
"vcmslcfg"}

\item{indicator}{A character, specifying the specific indicator of interest. Options are
"avg_rade9h", "avg_rade9h.masked", "cf_cvg" or "cvg"}

\item{link_base}{character, the base link to the EOG webpage/database}
}
\description{
Functions to download night time lights data from NASA's Earth Observation
Group's (EOG) database. The downloaded rasters are collected for different
time intervals, daily, monthly and annually.
}
