% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_utils.R
\name{get_terminal_executor}
\alias{get_terminal_executor}
\title{Get terminal executor for Ubuntu}
\usage{
get_terminal_executor(config)
}
\arguments{
\item{config}{The package configuration}
}
\value{
A character string with the terminal command to use
}
\description{
Determines the best terminal emulator to use on Ubuntu systems
}
\keyword{internal}
