% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_generalhelpers.R
\name{gunzip_reader}
\alias{gunzip_reader}
\title{A function to unzip a download}
\usage{
gunzip_reader(filename)
}
\arguments{
\item{filename}{the filenames of the downloads}
}
\description{
A function to unzip a download
}
