% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonalstats_plugin.R
\name{harmonize_weights}
\alias{harmonize_weights}
\title{A function to harmonize the weight raster with the raster to be extracted}
\usage{
harmonize_weights(raster_objs, weight_raster, method = "bilinear")
}
\arguments{
\item{raster_objs}{a list of rasters}

\item{weight_raster}{a list of weighting rasters of length 1 or the equal in length to
\code{raster_objs}}

\item{method}{see \code{terra::resample} for details}
}
\description{
This function will check and resample the weighting raster to match the data
raster which is to be extracted from. This step is necessary before zonal
statistics estimation that requires the use of weights
}
