% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hhgeo_dt}
\alias{hhgeo_dt}
\title{Nigeria GHS 2018/19 data}
\format{
## `hhgeo_dt`

An `data.table`, `data.frame` with 5116 rows and 42 variables:

\describe{
 \item{hhid}{Household ID}
 \item{dist_road2}{Distance to Nearest Road}
 \item{dist_popcenter2}{Distance to population center}
 ... see <https://microdata.worldbank.org/index.php/catalog/3557/related-materials>
 for the complete set of variables
 \item{geometry}{POINT location of each household}


}
}
\source{
<https://microdata.worldbank.org/index.php/catalog/3557>
}
\usage{
hhgeo_dt
}
\description{
The full household geovars dataset from Nigerian General Household Survey 2018/19
dataset.
}
\keyword{datasets}
