% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspire.R
\name{inspire}
\alias{inspire}
\alias{inspire_link}
\alias{is_inspire}
\alias{inspire_parse}
\title{INSPIRE geolinker}
\usage{
inspire_link(.data, id_col, x, y, res, crs = 3035, ...)

is_inspire(ids)

inspire_parse(ids)
}
\arguments{
\item{.data}{A dataframe with at least one column named according to the
\code{id_col} argument.}

\item{id_col}{Column in \code{.data} holding the spatial identifiers
in question.}

\item{x, y}{X and Y coordinates used to create the grid.}

\item{res}{Cell size of the grid in meters.}

\item{crs}{EPSG code of the CRS that \code{x} and \code{y} use.}

\item{...}{Not used}

\item{ids}{A character vector of IDs.}
}
\value{
\code{is_inspire}: A logical vector of length \code{length(ids)}.

\code{inspire_parse}: A list with fields \code{country} and \code{level}.

\code{inspire_link}: An sf dataframe containing \code{.data} and an added
geometry column.
}
\description{
Default geolinker registered through \code{\link{register_geolinker}}.
These functions are not intended to be used directly but through
\code{\link{enrich}}!

Contains functions to link, parse and validate grid geometries from
Eurostat's Infrastructure for Spatial Information in the
European Community (INSPIRE,
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/grids}).
}
\examples{
\dontshow{if () (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{# Link grid cells
enrich(c("CRS3035RES100mN1E2", "CRS3035RES100mN2E2"))}
\dontshow{\}) # examplesIf}
}
