% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lau.R
\name{lau}
\alias{lau}
\alias{lau_link}
\alias{is_lau}
\alias{lau_parse}
\title{LAU geolinker}
\usage{
lau_link(.data, id_col, country, ...)

is_lau(ids)

lau_parse(ids)
}
\arguments{
\item{.data}{A dataframe with at least one column named according to the
\code{id_col} argument.}

\item{id_col}{Column in \code{.data} holding the spatial identifiers
in question.}

\item{country}{Character vector of unique ISO-3 country codes contained
by \code{ids}. Only geometries within these countries are returned.}

\item{...}{Further arguments passed \code{\link[giscoR]{gisco_get_lau}}.}

\item{ids}{A character vector of IDs.}
}
\value{
\code{is_lau}: A logical vector of length \code{length(ids)}.

\code{lau_parse}: A list with fields \code{country} and \code{level}.

\code{lau_link}: An sf dataframe containing \code{.data} and an added
geometry column.
}
\description{
Default geolinker registered through \code{\link{register_geolinker}}.
These functions are not intended to be used directly but through
\code{\link{enrich}}!

Contains functions to link, parse and validate geometries from Eurostat's
Local Administrative Units (LAU,
\url{https://ec.europa.eu/eurostat/web/nuts/local-administrative-units}).

Requires the \href{https://ropengov.github.io/giscoR/}{\code{giscoR}}
package to be installed.
}
\examples{
\dontshow{if (requireNamespace("giscoR")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{# Link Normandy
enrich("FRD", linker = "nuts")}
\dontshow{\}) # examplesIf}
}
