% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgsd_utils.R
\name{list_from_rgsd}
\alias{list_from_rgsd}
\title{List files and folders from NINA's GeoSpatialData directory}
\usage{
list_from_rgsd(
  folder = "/data/R/GeoSpatialData",
  type = c("both", "file", "folder")
)
}
\arguments{
\item{folder}{Character. Either the full path or a relative subpath from \verb{/data/R/GeoSpatialData}.}

\item{type}{Character. One of \code{"both"} (default), \code{"file"}, or \code{"folder"} to filter results.}
}
\value{
A character vector of full paths, sorted with folders first.
}
\description{
Lists the contents of a specified folder within \verb{/data/R/GeoSpatialData} on NINA’s network.
}
