% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.r
\name{list_s3_geoparquet_files}
\alias{list_s3_geoparquet_files}
\title{List GeoParquet files in an S3 bucket (with optional prefix)}
\usage{
list_s3_geoparquet_files(
  bucket,
  prefix = NULL,
  endpoint = "s3-int-1.nina.no",
  use_https = TRUE,
  virtual = FALSE,
  full_info = TRUE
)
}
\arguments{
\item{bucket}{The S3 bucket name}

\item{prefix}{Optional prefix to filter files (e.g. "sam_env/")}

\item{endpoint}{The S3 endpoint (default is NINA's internal S3)}

\item{use_https}{Logical, use HTTPS for the connection (default TRUE)}

\item{virtual}{Logical, use virtual-hosted-style URLs (default FALSE)}

\item{full_info}{Logical, return a data.frame with file size and date (default TRUE)}
}
\value{
A character vector of file keys (or a data.frame if \code{full_info = TRUE})
}
\description{
List GeoParquet files in an S3 bucket (with optional prefix)
}
