% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naturalearth.R
\name{naturalearth}
\alias{naturalearth}
\alias{naturalearth_link}
\title{Natural Earth geolinker}
\usage{
naturalearth_link(.data, id_col, resolution = 50)
}
\arguments{
\item{.data}{A dataframe with at least one column named according to the
\code{id_col} argument.}

\item{id_col}{Column in \code{.data} holding the spatial identifiers
in question.}

\item{resolution}{Resolution of the polygon geometries. Can be 10
(1:10m), 50 (1:50m) or 110 (1:110m). A resolution of 1:10 requires
\href{https://docs.ropensci.org/rnaturalearthhires/}{\code{rnaturalearthhires}}
package to be installed. Otherwise, requires the
\href{https://docs.ropensci.org/rnaturalearthdata/}{\code{rnaturalearthdata}}
package. Note that at lower resolutions, some tiny countries disappear.}

\item{...}{Not used.}
}
\value{
An sf dataframe containing \code{.data} and an added geometry column.
}
\description{
Default geolinker registered through \code{\link{register_geolinker}}.
These functions are not intended to be used directly but through
\code{\link{enrich}}!

Contains functions to link  geometries from Natural Earth
(\url{https://www.naturalearthdata.com/}). Provides no functions for validating
and parsing as Natural Earth does not provide its own admin IDs.

Depending on the resolution selected, requires the \code{rnaturalearthdata}
or \code{rnaturalearthhires} packages to be installed.
}
\examples{
\donttest{# Monaco cannot be matched at resolution = 110
enrich("MCO", linker = "naturalearth", resolution = 110)

# But on higher resolutions, it can
enrich("MCO", linker = "naturalearth", resolution = 50)}
}
