% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_generalhelpers.R
\name{ntl_download_reader}
\alias{ntl_download_reader}
\title{A function to download and read in a file from the internet for Night Time lights}
\usage{
ntl_download_reader(
  url,
  ego_username,
  ego_password,
  client_id = "eogdata_oidc",
  client_secret = "2677ad81-521b-4869-8480-6d05b9e57d48",
  max_retries = 3,
  ...
)
}
\arguments{
\item{url}{a link URL}

\item{ego_username}{username for your eogdata account}

\item{ego_password}{password for your egodata account}

\item{...}{additional arguments to be used in `download.file()`}
}
\description{
A function to download and read in a file from the internet for Night Time lights
}
