% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_ntl_internals.R
\name{ntl_downloader}
\alias{ntl_downloader}
\title{An deprecated function used to download NTL rasters from the EOG website
This function was useful when EOG needed a username and password. Therefore
this is not really in use anymore unless the opportunity necessitates}
\usage{
ntl_downloader(
  client_id,
  client_secret,
  username,
  password,
  grant_type,
  token_url,
  data_url,
  shp_dt
)
}
\arguments{
\item{client_id}{the client ID}

\item{client_secret}{the client secret created on the website}

\item{username}{character, username for your eogdata account}

\item{password}{character, password for your egodata account}

\item{grant_type}{character, the grant type}

\item{token_url}{character, the token url}

\item{data_url}{character, the data url}

\item{shp_dt}{\code{sf}, \code{data.frame}, the shapefile object}
}
\description{
An deprecated function used to download NTL rasters from the EOG website
This function was useful when EOG needed a username and password. Therefore
this is not really in use anymore unless the opportunity necessitates
}
