% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts.R
\name{nuts}
\alias{nuts}
\alias{nuts_link}
\alias{is_nuts}
\alias{nuts_parse}
\title{NUTS geolinker}
\usage{
nuts_link(.data, id_col, country, level, ...)

is_nuts(ids)

nuts_parse(ids)
}
\arguments{
\item{.data}{A dataframe with at least one column named according to the
\code{id_col} argument.}

\item{id_col}{Column in \code{.data} holding the spatial identifiers
in question.}

\item{country}{Character vector of unique ISO-3 country codes contained
by \code{ids}. Only geometries within these countries are returned.}

\item{level}{Vector of geographic levels. Only geometries that adhere to
these levels are returned.}

\item{...}{Further arguments passed \code{\link[giscoR]{gisco_get_nuts}}.}

\item{ids}{A character vector of IDs.}
}
\value{
\code{is_nuts}: A logical vector of length \code{length(ids)}.

\code{nuts_parse}: A list with fields \code{country} and \code{level}.

\code{nuts_link}: An sf dataframe containing \code{.data} and an added
geometry column.
}
\description{
Default geolinker registered through \code{\link{register_geolinker}}.
These functions are not intended to be used directly but through
\code{\link{enrich}}!

Contains functions to link, parse and validate geometries from the
Nomenclature of Territorial Units for Statistics (NUTS,
\url{https://ec.europa.eu/eurostat/web/nuts/}).

Requires the \href{https://ropengov.github.io/giscoR/}{\code{giscoR}}
package to be installed.
}
\examples{
\dontshow{if (requireNamespace("giscoR")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{# Link Normandy
enrich("FRD", linker = "nuts")}
\dontshow{\}) # examplesIf}
}
