% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonalstats_plugin.R
\name{parallel_zonalstats}
\alias{parallel_zonalstats}
\title{Compute zonal statistics in-parallel}
\usage{
parallel_zonalstats(x, y, fun, numCores)
}
\arguments{
\item{x}{a \code{RasterLayer}, \code{RasterStack}, \code{RasterBrick}, or \code{SpatRaster} with data
to be extracted from}

\item{y}{a \code{sf}, \code{sfc}, \code{SpatialPolygonsDataFrame}, or \code{SpatialPolygons} object
with polygonal geometries to be extracted into}

\item{fun}{the function to be used compute zonal statistics}

\item{numCores}{the number of cores to be used in parallel processing}
}
\description{
A function to split a shapefile into many parts and computes the zonal statistics
for each part in parallel to speed up the zonal statistics estimations
}
