% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertile_processor.R
\name{point_sf_to_raster}
\alias{point_sf_to_raster}
\title{A function convert point data to raster}
\usage{
point_sf_to_raster(point_sf, crs = "EPSG:4326", resolution, agg_fun = sum)
}
\arguments{
\item{point_sf}{`sf`, `data.frame` object with point geometries, the first
column is assumed contain the intended raster's field value.}

\item{crs}{integer or crs projection string, default is EPSG:4326}

\item{resolution}{numeric, the value of the raster's intended resolution}

\item{agg_fun}{the aggregation function (e.g. sum, mean etc)}
}
\description{
A function convert point data to raster
}
\details{
Ensure, the `sf`, `data.frame` object contains the field value first and
then the geometry and `resolution` of `point_dt` is in the same units as `crs`.
i.e. a degree resolution only works correctly if `point_dt` has a `crs` in
degrees and likewise for a metric resolution. In addition, if the point_dt
comes as a `data.frame` object with latitude and longitude, it must be
converted into an `sf`, `data.frame` by using the `sf::st_as_sf()` function.
See help(sf::st_as_sf) for more details on the conversion process. In
specifying the `crs`, simply using the 4 digit number will provide an error.
See default for example.
}
\examples{
\donttest{

 #example usage with shapefile
 raster_obj <- point_sf_to_raster(point_sf = point_dt,
                                  crs = "EPSG:4326",
                                  resolution = 0.1)

}
}
