% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popHDX_dt}
\alias{popHDX_dt}
\title{Population data for Nigeria}
\format{
A \code{data.frame} containing 773 areas with identifiers and total population.
\describe{
\item{ADM2_EN}{Admin 2 division names, i.e. Local Government Area names}
\item{ADM2_PCODE}{Admin 2 codes, i.e. Local Government Area codes}
\item{T_TL}{Total population for area}
\item{...}{For more information see \url{https://data.humdata.org/dataset/cod-ps-nga}}
}
}
\source{
Humanitarian Data Exchange
}
\usage{
popHDX_dt
}
\description{
Population data for second administrative divisions (local government areas.)
}
\keyword{datasets}
