% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode.R
\name{postcode}
\alias{postcode}
\alias{postcode_link}
\title{Postcode geolinker}
\usage{
postcode_link(.data, id_col, username, ...)
}
\arguments{
\item{.data}{A dataframe with at least one column named according to the
\code{id_col} argument.}

\item{id_col}{Column in \code{.data} holding the spatial identifiers
in question.}

\item{username}{Username for the GeoNames API. Necessary for any request.
For details, see the
\href{https://www.geonames.org/export/web-services.html}{documentation}}

\item{...}{Further arguments passed to the \code{postalCodeSearch} endpoint
of GeoNames. For details, see the
\href{https://www.geonames.org/export/web-services.html}{documentation}.}
}
\value{
An sf dataframe containing \code{.data} and an added geometry column.
}
\description{
Default geolinker registered through \code{\link{register_geolinker}}.
These functions are not intended to be used directly but through
\code{\link{enrich}}!

Contains functions to link point geometries from GeoNames' postal codes
database (\url{https://www.geonames.org/postal-codes/}). Provides no functions for
validating and parsing as postal codes are usually pretty generic numbers.

Requires the \href{https://docs.ropensci.org/geonames/}{\code{geonames}}
package to be installed.
}
\examples{
\donttest{try(enrich("50667", linker = "postcode"))}
}
