% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonalstats_plugin.R
\name{postdownload_processor}
\alias{postdownload_processor}
\title{A function to process downloaded raster into the final output in shapefile/geocoded survey}
\usage{
postdownload_processor(
  raster_objs,
  survey_dt = NULL,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  extract_fun = "mean",
  buffer_size = NULL,
  survey_crs = 4326,
  name_set,
  shp_dt = NULL,
  shp_fn = NULL,
  grid_size = 1000,
  return_raster,
  weight_raster,
  zonalstats_fun = compute_zonalstats
)
}
\arguments{
\item{raster_objs}{a list of rasters}

\item{survey_dt}{a sf/data.frame object, the geocoded survey}

\item{survey_fn}{a character, the filename of the survey (.dta file) (STATA users only)}

\item{survey_lat}{a character, for latitude variable name in survey_fn
(STATA users only)}

\item{survey_lon}{a character, for longitude variable name in survey_fn
(STATA users only)}

\item{extract_fun}{a function string to compute the extract, i.e. mean, min, max,
weighted_mean, quantiles, etc...}

\item{buffer_size}{a numeric, radius of buffer for survey_dt (or survey_fn) units.}

\item{survey_crs}{an integer or character, the CRS for the survey}

\item{name_set}{a character vector of the names for the extracted raster
values}

\item{shp_dt}{an sf object with polygons}

\item{shp_fn}{a character, the name of shapefile for STATA users only (.shp/.gpkg file)}

\item{grid_size}{a numeric, size of the size of the grid in meters}

\item{weight_raster}{(optional) a raster to be used as weights for weighted
statistics}

\item{zonalstats_fun}{the function that extracts the rasters into the polygons, called
with the arguments of \code{compute_zonalstats} (the default). \code{geolink_buildings} uses it to
choose the aggregation per layer.}
}
\description{
A function to process downloaded raster into the final output in shapefile/geocoded survey
}
