% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoareacreator_plugin.R
\name{prep_sf_obj_predownload}
\alias{prep_sf_obj_predownload}
\title{A quick function to prepare shapefile objects prior to downloading spatial data}
\usage{
prep_sf_obj_predownload(
  shp_dt = NULL,
  survey_dt = NULL,
  shp_fn = NULL,
  survey_fn = NULL
)
}
\arguments{
\item{shp_dt}{the shapefile object \verb{sf data.frame}}

\item{survey_dt}{the survey object \verb{sf data.frame}}

\item{shp_fn}{character, the shapefile filename path ending in .shp or json}

\item{survey_fn}{character, the survey filename path ending in .dta}
}
\description{
A quick function to prepare shapefile objects prior to downloading spatial data
}
