% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_landcover.R
\name{process_raster}
\alias{process_raster}
\title{Process a single raster (project to 4326)}
\usage{
process_raster(path, output_path)
}
\arguments{
\item{path}{Path to input raster}

\item{output_path}{Path to save processed raster}
}
\value{
Path to processed raster or NULL if processing failed
}
\description{
Process a single raster (project to 4326)
}
