% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_landcover.R
\name{process_year_data}
\alias{process_year_data}
\title{Process year data}
\usage{
process_year_data(year, sf_obj, session_temp_dir, resample_resolution = NULL)
}
\arguments{
\item{year}{Year to process}

\item{sf_obj}{Spatial object}

\item{session_temp_dir}{Temporary directory}

\item{resample_resolution}{Resolution for resampling}
}
\value{
SF object with results for this year
}
\description{
Process year data
}
