% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_utils.R
\name{read_config}
\alias{read_config}
\title{Read package configuration}
\usage{
read_config()
}
\value{
A list containing configuration settings
}
\description{
Reads the configuration file for the package based on the detected operating system
}
\keyword{internal}
