% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_utils.r
\name{read_geoparquet_from_s3}
\alias{read_geoparquet_from_s3}
\title{Read GeoParquet spatial data from S3}
\usage{
read_geoparquet_from_s3(
  s3_path,
  bucket,
  endpoint = "s3-int-1.nina.no",
  use_https = TRUE,
  virtual = FALSE,
  ...
)
}
\arguments{
\item{s3_path}{The object key (path to the file in the bucket)}

\item{bucket}{The S3 bucket name}

\item{endpoint}{The S3 endpoint (default is NINA's internal S3)}

\item{use_https}{Logical, use HTTPS for the connection (default TRUE)}

\item{virtual}{Logical, use virtual-hosted-style URLs (default FALSE)}

\item{...}{Additional arguments passed to \code{sfarrow::st_read_parquet()}}
}
\value{
An \code{sf} object read from S3
}
\description{
Read GeoParquet spatial data from S3
}
