% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_generalhelpers.R
\name{read_survey_data}
\alias{read_survey_data}
\title{Helper function to read survey data}
\usage{
read_survey_data(file_path)
}
\arguments{
\item{file_path}{character, a file location path}
}
\description{
Helper function to read survey data
}
