% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr.R
\name{unhcr}
\alias{unhcr}
\alias{unhcr_link}
\alias{is_unhcr}
\alias{unhcr_parse}
\title{UNHCR geolinker}
\usage{
unhcr_link(
  .data,
  id_col,
  country = NULL,
  level = 0,
  resolution = c("1", "15", "25"),
  ...
)

is_unhcr(ids)

unhcr_parse(ids)
}
\arguments{
\item{.data}{A dataframe with at least one column named according to the
\code{id_col} argument.}

\item{id_col}{Column in \code{.data} holding the spatial identifiers
in question.}

\item{country}{Character vector of unique ISO-3 country codes contained
by \code{ids}. Only geometries within these countries are returned.}

\item{level}{Vector of geographic levels. Only geometries that adhere to
these levels are returned. Can be an integer between 0 (country level)
and 2 (regional level).}

\item{resolution}{Resolution of the polygon geometries. Can be 1
(1:1,000,000), 15 (1:15,000,000) or 25 (1:25,000,000). Ignored if
\code{level != 0}.}

\item{...}{Further arguments passed to the query endpoint of the
UNHCR API. See the \href{https://gis.unhcr.org/arcgis/sdk/rest/index.html#/Query_Feature_Service_Layer/02ss0000002r000000/}{documentation}
for details.}

\item{ids}{A character vector of IDs.}
}
\value{
\code{is_unhcr}: A logical vector of length \code{length(ids)}.

\code{unhcr_parse}: A list with fields \code{country} and \code{level}.

\code{unhcr_link}: An sf dataframe containing \code{.data} and an added
geometry column.
}
\description{
Default geolinker registered through \code{\link{register_geolinker}}.
These functions are not intended to be used directly but through
\code{\link{enrich}}!

Contains functions to link, parse and validate geometries from the Office
of the United Nations High Commissioner for Refugees (UNHCR,
\url{https://data.unhcr.org/en/geoservices/}).
}
\examples{
\donttest{enrich("MCO", linker = "unhcr")}
}
