% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_ntl_internals.R
\name{valid_url}
\alias{valid_url}
\title{Check if the url actually works and contains data}
\usage{
valid_url(url, t = 2)
}
\arguments{
\item{url}{character, the url}
}
\description{
Check if the url actually works and contains data
}
