% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verify_python_env}
\alias{verify_python_env}
\title{Verify Python Virtual Environment}
\usage{
verify_python_env()
}
\value{
A list containing virtual environment details
}
\description{
Checks the status of the Python virtual environment for the package
}
