% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonalstats_plugin.R
\name{zonalstats_prepshp}
\alias{zonalstats_prepshp}
\title{A function to process raster downloads and compute zonal statistics for shapefiles
and geocoded surveys}
\usage{
zonalstats_prepshp(shp_dt, shp_fn = NULL, grid_size)
}
\arguments{
\item{shp_dt}{a sf/data.frame object, the shapefile}

\item{shp_fn}{a character, the name of shapefile for STATA users only (.shp/.gpkg file)}

\item{grid_size}{a numeric, size of the size of the grid in meters}

\item{varname_tag}{a character, a tag on to be used in created the set of variables
the common portion of the variable name (e.g. "ntl_" for the night time light
variables)}

\item{time_unit}{a character, for the time intervals for the dataset e.g. daily,
monthly, annual}

\item{name_count}{a character, the number of periods of data collected e.g. 10 days,
vs or 10 months or 10 years etc.}

\item{raster_objs}{a list of raster object(s), a list of raster objects}

\item{survey_fn}{a character, the filename of the survey (.dta file) (STATA users only)}

\item{survey_lat}{a character, for latitude variable name in survey_fn
(STATA users only)}

\item{survey_lon}{a character, for longitude variable name in survey_fn
(STATA users only)}

\item{survey_dt}{a sf/data.frame object, the geocoded survey}

\item{extract_fun}{a character, an extraction fun. See `exactextractr::exact_extract()`
for examples}

\item{buffer_size}{a numeric, radius of buffer for survey_dt (or survey_fn) units.}
}
\description{
A function to process raster downloads and compute zonal statistics for shapefiles
and geocoded surveys
}
