% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonalstats_plugin.R
\name{zonalstats_prepsurvey}
\alias{zonalstats_prepsurvey}
\title{A function to process raster downloads and compute zonal statistics for shapefiles
and geocoded surveys}
\usage{
zonalstats_prepsurvey(
  survey_dt,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  buffer_size = NULL,
  survey_crs = NULL
)
}
\arguments{
\item{survey_dt}{a sf/data.frame object, the geocoded survey}

\item{survey_fn}{a character, the filename of the survey (.dta file) (STATA users only)}

\item{survey_lat}{a character, for latitude variable name in survey_fn
(STATA users only)}

\item{survey_lon}{a character, for longitude variable name in survey_fn
(STATA users only)}

\item{buffer_size}{a numeric, radius of buffer for survey_dt (or survey_fn) units.}

\item{survey_crs}{an integer or character, the CRS for the survey}
}
\description{
A function to process raster downloads and compute zonal statistics for shapefiles
and geocoded surveys
}
