% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastertile_processor.R
\name{as.data.table.raster}
\alias{as.data.table.raster}
\title{Convert a raster to data.table object}
\usage{
\method{as.data.table}{raster}(
  x,
  keep.rownames = FALSE,
  xy = FALSE,
  inmem = raster::canProcessInMemory(x, 2),
  ...
)
}
\arguments{
\item{x}{An object of class \verb{Raster*}, typically from the \code{raster} package.}

\item{keep.rownames}{Logical. If \code{TRUE}, include row names as a column in the output.}

\item{xy}{(optional) Logical. If \code{TRUE}, include coordinates of raster cells in the output.}

\item{inmem}{(optional) Logical. If \code{TRUE}, process the raster in memory. Default is determined by \code{raster::canProcessInMemory()}.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Convert a raster to data.table object
}
