% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonalstats_plugin.R
\name{compute_zonalstats}
\alias{compute_zonalstats}
\title{A function to extract raster values into the shapefile}
\usage{
compute_zonalstats(
  shp_dt,
  raster_objs,
  extract_fun,
  name_set,
  weight_raster = NULL
)
}
\arguments{
\item{shp_dt}{an sf object with polygons}

\item{raster_objs}{a list of rasters}

\item{extract_fun}{a function string to compute the extract, i.e. mean, min, max,
weighted_mean, quantiles, etc...}

\item{name_set}{a character vector of the names for the extracted raster
values}

\item{weight_raster}{(optional) a raster to be used as weights for weighted
statistics}
}
\description{
This function will extract raster values as zonal statistics into the
corresponding polygons in the shapefile
}
