% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_buildings}
\alias{geolink_buildings}
\title{Download high resolution building data from WorldPop}
\usage{
geolink_buildings(
  version,
  iso_code,
  shp_dt = NULL,
  shp_fn = NULL,
  grid_size = NULL,
  survey_dt = NULL,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  buffer_size = NULL,
  extract_fun = "mean",
  survey_crs = 4326,
  indicators = "ALL"
)
}
\arguments{
\item{version}{A character, the version of the building data to download. Options are "v1.1" and "v2.0".}

\item{iso_code}{A character, the ISO country code for the country of interest.}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons representing the study area.}

\item{shp_fn}{A character, file path for the shapefile (.shp) to be read (for STATA users only).}

\item{grid_size}{A numeric, the grid size to be used in meters for analyzing the building data.}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey with latitude and longitude values (optional).}

\item{survey_fn}{A character, file path for geocoded survey (.dta format) (for STATA users only & if use_survey is TRUE) (optional).}

\item{survey_lat}{A character, latitude variable from survey (for STATA users only & if use_survey is TRUE) (optional).}

\item{survey_lon}{A character, longitude variable from survey (for STATA users only & if use survey is TRUE) (optional).}

\item{buffer_size}{A numeric, the buffer size to be used around each point in the survey data, in meters (optional).}

\item{extract_fun}{A character, a function to be applied in extraction of raster into the shapefile.
Default is "mean". Other options are "sum", "min", "max", "sd", "skew" and "rms" (optional).}

\item{survey_crs}{An integer, the Coordinate Reference System (CRS) for the survey data. Default is 4326 (WGS84) (optional).}
}
\value{
A processed data frame or object based on the input parameters and downloaded data.
}
\description{
This function downloads high-resolution building data from WorldPop based on the specified version and ISO country code. It can incorporate survey data for further analysis. The building data is downloaded as raster files and can be processed and analyzed using the `postdownload_processor` function.
}
\examples{
\donttest{

# Example usage with version 1.1
df <- geolink_buildings(version = "v1.1", iso_code = "NGA", shp_dt = shp_dt[shp_dt$ADM1_PCODE == "NG001",])

}

}
