% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_CMIP6}
\alias{geolink_CMIP6}
\title{Download CMIP6 climate model data}
\usage{
geolink_CMIP6(
  start_date,
  end_date,
  scenario,
  desired_models,
  shp_dt = NULL,
  shp_fn = NULL,
  grid_size = NULL,
  survey_dt = NULL,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  buffer_size = NULL,
  extract_fun = "mean",
  survey_crs = 4326,
  return_raster = FALSE,
  weight_raster = NULL
)
}
\arguments{
\item{start_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{end_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{scenario}{A scenario has to be selected and can be one of "historical", "ssp245" or "ssp585"}

\item{desired_models}{The name or names in a list, of the desired model(s) required for the analysis, for example
("ACCESS-CM-2" or c("ACCESS-CM-2","UKESM1-0-LL")). See the cmip6:model summary in the STAC collection for a full list of models.}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons representing the study area.}

\item{shp_fn}{A character, file path for the shapefile (.shp) to be read (for STATA users only).}

\item{grid_size}{A numeric, the grid size to be used in meters for analyzing the climate model data.}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey with latitude and longitude values (optional).}

\item{survey_fn}{A character, file path for geocoded survey (.dta format) (for STATA users only & if use_survey is TRUE) (optional).}

\item{survey_lat}{A character, latitude variable from survey (for STATA users only & if use_survey is TRUE) (optional).}

\item{survey_lon}{A character, longitude variable from survey (for STATA users only & if use survey is TRUE) (optional).}

\item{buffer_size}{A numeric, the buffer size to be used around each point in the survey data, in meters (optional).}

\item{extract_fun}{A character, a function to be applied in extraction of raster into the shapefile.
Default is "mean". Other options are "sum", "min", "max", "sd", "skew" and "rms" (optional).}

\item{survey_crs}{An integer, the Coordinate Reference System (CRS) for the survey data. Default is 4326 (WGS84) (optional).}

\item{return_raster}{logical, default is FALSE, if TRUE a raster will be returned ONLY. The resulting
raster is cropped to the extent of \code{shp_dt} if \code{shp_dt} is specified. Otherwise, full raster from
source is downloaded.}

\item{weight_raster}{a raster object of class \code{spatRaster} or a list of \code{spatRaster} objects}
}
\value{
A processed data frame based on the input parameters and downloaded data.
}
\description{
Download CMIP6 climate model data
}
\details{
This function downloads and processes CMIP6 (Coupled Model Intercomparison Project Phase 6)
climate model data for a specific scenario and desired model.
It allows for further analysis of the data in conjunction with geographic data.
Please see x for the full list of possible models and scenarios which can be downloaded.
The data is extracted into a shapefile provided by user. An added service for tesselating/gridding
the shapefile is also provided for users that need this data further analytics that require
equal area zonal statistics. Shapefile estimates at the grid or native polygon level is a
permitted final output. However, a geocoded survey with population estimates are the end goal
if the user also chooses. The function will merge shapefile polygons (either gridded or
native polygons) with the location of the survey units i.e. population estimates for the
locations of the units within the survey will be returned. The function is also set up for
stata users and allows the user to pass file paths for the household survey \code{survey_fn}
(with the lat and long variable names \code{survey_lon} and \code{survey_lat}) as well. This is requires
a .dta file which is read in with \code{haven::read_dta()} package. Likewise, the user is permitted
to pass a filepath for the location of the shapefile \code{shp_fn} which is read in with the
\code{sf::read_sf()} function.
}
\examples{

\dontrun{
\donttest{

 #example usage
df <- geolink_CMIP6(start_date = "2019-01-01",
                    end_date = "2019-12-31",
                    scenario = "ssp245",
                   desired_models = "UKESM1-0-LL",
                   shp_dt = shp_dt[shp_dt$ADM1_EN == "Abia",],
                   grid_size = 1000)
}}

}
