% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_electaccess}
\alias{geolink_electaccess}
\title{Download high resolution electrification access data from HREA}
\usage{
geolink_electaccess(
  start_date = NULL,
  end_date = NULL,
  shp_dt = NULL,
  shp_fn = NULL,
  grid_size = 1000,
  survey_dt = NULL,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  buffer_size = NULL,
  extract_fun = "mean",
  survey_crs = 4326
)
}
\arguments{
\item{start_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{end_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons}

\item{shp_fn}{A character, file path for the shapefile (.shp) to be read (for STATA users only)}

\item{grid_size}{A numeric, the grid size to be used in meters}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey i.e. a household survey with latitude and longitude values.}

\item{survey_fn}{A character, file path for geocoded survey (.dta format) (for STATA users only & if use_survey is TRUE)}

\item{survey_lat}{A character, latitude variable from survey (for STATA users only & if use_survey is TRUE)}

\item{survey_lon}{A character, longitude variable from survey (for STATA users only & if use survey is TRUE)}

\item{buffer_size}{A numeric, the size of the buffer for \code{survey_dt} or \code{survey_fn} in meters.}

\item{extract_fun}{A character, a function to be applied in extraction of raster into the shapefile.
Default is mean. Other options are "sum", "min", "max", "sd", "skew" and "rms".}

\item{survey_crs}{A numeric, the default is 4326}
}
\description{
Download high resolution electrification access data from HREA
}
\details{
This function downloads and processes the following electrification access indicators, lightscore,
light-composite, nightime proportion and estimated brightness.
Details for the dataset can be found here: https://hrea.isr.umich.edu/.
The data is extracted into a shapefile provided by user. An added service for tesselating/gridding
the shapefile is also provided for users that need this data further analytics that require
equal area zonal statistics. Shapefile estimates at the grid or native polygon level is a
permitted final output. However, a geocoded survey with population estimates are the end goal
if the user also chooses. The function will merge shapefile polygons (either gridded or
native polygons) with the location of the survey units i.e. population estimates for the
locations of the units within the survey will be returned. The function is also set up for
stata users and allows the user to pass file paths for the household survey \code{survey_fn}
(with the lat and long variable names \code{survey_lon} and \code{survey_lat}) as well. This is requires
a .dta file which is read in with \code{haven::read_dta()} package. Likewise, the user is permitted
to pass a filepath for the location of the shapefile \code{shp_fn} which is read in with the
\code{sf::read_sf()} function.
}
\examples{
\donttest{


df = geolink_electaccess(shp_dt = shp_dt[shp_dt$ADM1_EN == "Abia",],
  start_date = "2018-12-31", end_date = "2019-12-31")

elect_shp_fn = geolink_electaccess(shp_fn = "tests/testthat/testdata/shp_dt.shp",
start_date = "2018-12-31",
end_date = "2019-12-31")

elect_survey <- geolink_electaccess(
 start_date = "2019-01-01",
 end_date = "2019-12-31",
 survey_dt = hhgeo_dt[hhgeo_dt$ADM1_EN == "Abia",],
 buffer_size = 1000)

elect_survey_fn <- geolink_electaccess(start_date = "2019-01-01",
                                      end_date = "2019-12-31",
                                      survey_fn = "tests/testthat/testdata/xy_hhgeo_dt.dta",
                                      survey_lon = "x",
                                      survey_lat = "y",
                                      buffer_size = 1000)
}

}
