% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_get_poi}
\alias{geolink_get_poi}
\title{Download points of interest from OSM data using open street maps API.}
\usage{
geolink_get_poi(
  osm_key,
  osm_value = NULL,
  shp_dt = NULL,
  shp_fn = NULL,
  survey_dt = NULL,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  buffer_size = NULL,
  survey_crs = 4326,
  grid_size = NULL
)
}
\arguments{
\item{osm_key}{A character, refering to the osm key wiki page, please see details below}

\item{osm_value}{A character, refering to the osm key wiki page, please see details below}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons}

\item{shp_fn}{A link to location of shapefile for stat users}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey i.e. a household survey with latitude and longitude values.}

\item{survey_fn}{A character, file path for geocoded survey (.dta format) (for STATA users only & if use_survey is TRUE)}

\item{survey_lat}{A character, latitude variable from survey (for STATA users only & if use_survey is TRUE) (optional).}

\item{survey_lon}{A character, longitude variable from survey (for STATA users only & if use survey is TRUE) (optional).}

\item{buffer_size}{buffer area around shapefile or survey points.}

\item{survey_crs}{An integer, the Coordinate Reference System (CRS) for the survey data. Default is 4326 (WGS84) (optional).}

\item{grid_size}{A numeric, the grid size to be used as a buffer around survey points.}
}
\description{
Download points of interest from OSM data using open street maps API.
}
\details{
This function downloads open street maps (osm) datapoints based on osn_key and osm_value
arguments passed to the function. The full details for osm_key and osm_value arguments
can be found here: https://wiki.openstreetmap.org/wiki/Map_features. Either a shapefile, shapefile path,
household survey or household survey path can be passed in as an area for analysis.
The function works for shapefiles, however is also set up for stata users and allows the user to
pass file paths for the household survey \code{survey_fn}
(with the lat and long variable names \code{survey_lon} and \code{survey_lat}) as well. This is requires
a .dta file which is read in with \code{haven::read_dta()} package. Likewise, the user is permitted
to pass a filepath for the location of the shapefile \code{shp_fn} which is read in with the
\code{sf::read_sf()} function.
}
\examples{

\dontrun{
\donttest{


poi_survey_df <- geolink_get_poi(osm_key = "amenity",
                                buffer_size = 2000,
                                survey_dt = hhgeo_dt[hhgeo_dt$ADM1_EN == "Abia",])


poi_survey_fn <- geolink_get_poi(
 osm_key = "amenity",
 buffer_size = 2000,
 survey_fn = "tests/testthat/testdata/xy_hhgeo_dt.dta",
 survey_lon = "x",
 survey_lat = "y",
 survey_crs = 4326)


poi_shp = geolink_get_poi(osm_key = "amenity",
                         shp_dt = shp_dt[shp_dt$ADM1_EN == "Abia",])

poi_shp_fn = geolink_get_poi(osm_key = "amenity",
                             shp_fn = "tests/testthat/testdata/shp_dt.shp")


}}
}
