% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_landcover}
\alias{geolink_landcover}
\title{Download and Merge Annual Land Use Land Cover data into geocoded surveys}
\usage{
geolink_landcover(
  start_date = NULL,
  end_date = NULL,
  shp_dt = NULL,
  survey_dt = NULL
)
}
\arguments{
\item{start_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{end_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey i.e.}
}
\description{
Download Land Use Land Cover data from the LULC dataset at annual intervals for a specified period
The data is downloaded in raster format and combined with shapefile and/or survey data provided
by the user. Source data: https://planetarycomputer.microsoft.com/dataset/io-lulc-annual-v02
}
\examples{

\donttest{

#loading the survey data and shapefile

data("hhgeo_dt")
data("shp_dt")

#pull annual land use land cover and combine with household survey based on
#grid tesselation of shapefile at 1000m

df <- geolink_landcover(


}

}
