% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_landcover}
\alias{geolink_landcover}
\title{Download and Merge Annual Land Use Land Cover data into geocoded surveys}
\usage{
geolink_landcover(
  start_date,
  end_date,
  shp_dt = NULL,
  shp_fn = NULL,
  survey_dt = NULL,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  buffer_size = NULL,
  survey_crs = 4326,
  grid_size = NULL,
  use_resampling = TRUE,
  target_resolution = 1000,
  return_raster = FALSE,
  weight_raster = NULL
)
}
\arguments{
\item{start_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{end_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons}

\item{shp_fn}{Character string. Path to a shapefile if shp_dt is not provided directly}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey}

\item{survey_fn}{Character string. Path to survey data file (.dta or .csv)}

\item{survey_lat}{Character string. Column name for latitude in survey data}

\item{survey_lon}{Character string. Column name for longitude in survey data}

\item{buffer_size}{Numeric. Buffer size (in meters) to apply around survey points}

\item{survey_crs}{Numeric. CRS code for survey coordinates (default: 4326)}

\item{grid_size}{Numeric. Size of grid cells for tessellation (in meters)}

\item{use_resampling}{Logical. Whether to resample rasters to a common resolution (default: TRUE)
Using resampling might use a significant amount of memory.}

\item{target_resolution}{Numeric. Target resolution for resampling in meters (default: 1000)}

\item{return_raster}{logical, default is FALSE, if TRUE a raster will be returned ONLY. The resulting
raster is cropped to the extent of \code{shp_dt} if \code{shp_dt} is specified. Otherwise, full raster from
source is downloaded.}

\item{weight_raster}{a raster object of class \code{spatRaster} or a list of \code{spatRaster} objects}
}
\value{
An sf object with land cover classifications by year
}
\description{
Download Land Use Land Cover data from the LULC dataset at annual intervals for a specified period.
The data is downloaded in raster format and combined with shapefile and/or survey data provided
by the user. Source data: https://planetarycomputer.microsoft.com/dataset/io-lulc-annual-v02
}
\examples{
\dontrun{
\donttest{
# Loading the survey data and shapefile
data("hhgeo_dt")
data("shp_dt")

# Pull annual land use land cover for Abia region

 landcover_shp <- geolink_landcover(
 start_date = "2019-01-01",
 end_date = "2019-12-31",
 shp_dt = shp_dt[shp_dt$ADM1_EN == "Abia",],
 use_resampling = TRUE,
 target_resolution = 1000)



 landcover_survey <- geolink_landcover(
 start_date = "2020-01-01",
 end_date = "2020-12-31",
 survey_dt = hhgeo_dt[hhgeo_dt$ADM1_EN == "Abia",],
 buffer_size = 1000,
 use_resampling = TRUE,
 target_resolution = 1000)


}
}
}
