% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_ntl}
\alias{geolink_ntl}
\title{Download and Merge Monthly and Annual Night Time Light data into geocoded surveys}
\usage{
geolink_ntl(
  time_unit = "annual",
  start_date,
  end_date,
  version,
  indicator,
  slc_type,
  shp_dt,
  shp_fn = NULL,
  grid = FALSE,
  grid_size = 1000,
  use_survey = TRUE,
  survey_dt,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL
)
}
\arguments{
\item{time_unit}{A character, either "month" or "annual" monthly or annual rainfall aggregates
are to be estimated}

\item{start_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{end_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{version}{A string character, for now only "v10" is supported more to come later}

\item{indicator}{A character, specifying the specific indicator of interest. Options are
"average", "average_masked", "cf_cvg", "cvg", "lit_mask", "maximum", "median", "median_masked" and "minimum"}

\item{slc_type}{A character, adjustment type made to the NTL data either "vcmcfg" or
"vcmslcfg"}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons}

\item{shp_fn}{A character, file path for the shapefile (.shp) to be read (for STATA users only)}

\item{grid}{logical, TRUE if shapefile needs to tesselated/gridded before estimating grid level
zonal statistics and performing a spatial join to the household survey}

\item{grid_size}{A numeric, the grid size to be used in meters}

\item{use_survey}{logical, if TRUE a survey is expected (i.e. survey_dt cannot be null)}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey i.e.
a household survey with latitude and longitude values.}

\item{survey_fn}{A character, file path for geocoded survey (.dta format) (for STATA users only &
if use_survey is TRUE)}

\item{survey_lat}{A character, latitude variable from survey (for STATA users only &
if use_survey is TRUE)}

\item{survey_lon}{A character, longitude variable from survey (for STATA users only &
if use survey is TRUE)}

\item{extract_fun}{A character, a function to be applied in extraction of raster into the shapefile.
Default is mean. Other options are "sum", "min", "max", "sd", "skew" and "rms".}
}
\description{
Download rainfall data from the CHIRPS data at monthly/annual intervals for a specified period
The data is downloaded in raster format and combined with shapefile and/or survey data provided
by the user
}
\details{
NTL data is sourced from the NASA's Earth Observation Group database.
The data is extracted into a shapefile provided by user. An added service for tesselating/gridding
the shapefile is also provided for users that need this data further analytics that require
equal area zonal statistics. Shapefile estimates at the grid or native polygon level is a
permitted final output. However, a geocoded survey with rainfall estimates are the end goal
if the user also chooses. The function will merge shapefile polygons (either gridded or
native polygons) with the location of the survey units i.e. rainfall estimates for the
locations of the units within the survey will be returned. The function is also set up for
stata users and allows the user to pass file paths for the household survey `survey_fn`
(with the lat and long variable names `survey_lon` and `survey_lat`) as well. This is requires
a .dta file which is read in with `haven::read_dta()` package. Likewise, the user is permitted
to pass a filepath for the location of the shapefile `shp_fn` which is read in with the
`sf::read_sf()` function.
}
