% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_ntl}
\alias{geolink_ntl}
\title{Download and Merge Monthly and Annual Night Time Light data into geocoded surveys}
\usage{
geolink_ntl(
  time_unit = "annual",
  start_date,
  end_date,
  annual_version = "v21",
  month_version = "v10",
  indicator,
  slc_type = "vcmslcfg",
  shp_dt,
  shp_fn = NULL,
  grid_size = NULL,
  survey_dt,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  extract_fun = "mean",
  buffer_size = NULL,
  survey_crs = 4326
)
}
\arguments{
\item{time_unit}{A character, either "month" or "annual" monthly or annual rainfall aggregates
are to be estimated}

\item{start_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{end_date}{An object of class date, must be specified like "yyyy-mm-dd"}

\item{annual_version}{A character, the version of annual EOG data for download, default set to "v21"}

\item{month_version}{A character, the version of month EOG data to use. default set to "v10".}

\item{indicator}{A character, specifying the specific indicator of interest. Options are
"average", "average_masked", "cf_cvg", "cvg", "lit_mask", "maximum", "median",
"median_masked" and "minimum"}

\item{slc_type}{A character, adjustment type made to the NTL data either "vcmcfg" or
"vcmslcfg"}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons}

\item{shp_fn}{A character, file path for the shapefile (.shp) to be read (for STATA users only)}

\item{grid_size}{A numeric, the grid size to be used in meters. Defaults to NULL.}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey i.e.
a household survey with latitude and longitude values.}

\item{survey_fn}{A character, file path for geocoded survey (.dta format) (for STATA users only)}

\item{survey_lat}{A character, latitude variable from survey (for STATA users only &
if use_survey is TRUE)}

\item{survey_lon}{A character, longitude variable from survey (for STATA users only &
if use survey is TRUE)}

\item{extract_fun}{A character, a function to be applied in extraction of raster into the shapefile.
Default is mean. Other options are "sum", "min", "max", "sd", "skew" and "rms".}

\item{buffer_size}{A numeric, the size of the buffer for `survey_dt` or `survey_fn` in meters.}

\item{survey_crs}{A numeric, the default is 4326}
}
\description{
Download night lights luminosity at monthly/annual intervals for a specified period
The data is downloaded in raster format and combined with shapefile and/or survey data provided
by the user
}
\details{
NTL data is sourced from the NASA's Earth Observation Group database.
The data is extracted into `shp_dt` shapefile object. An added service for tesselating/gridding
the `shp_dt` is also provided if `grid_size` is specified for further analytics that require
equal specified area zonal statistics. Shapefile estimates at the grid or native polygon level is a
permitted final output. Also, if `survey_dt` is not NULL, the user may compute luminosity within
neighborhood of a household given a specified `buffer_size`. The function draw polygons around
household locations of `buffer_size` with zonal statistics computed as formerly described.
The function is also set up for stata users and allows the user to pass file paths for the
household survey `survey_fn` (with the lat and long variable names `survey_lon` and `survey_lat`)
as well. This is requires a .dta file which is read in with `haven::read_dta()` package.
Likewise, the user is permitted to pass a filepath for the location of the shapefile `shp_fn`
which is read in with the `sf::read_sf()` function.
}
\examples{

\donttest{

#loading the survey data and shapefile

data("hhgeo_dt")
data("shp_dt")

#pull monthly night lights and combine with household survey based on
#grid tesselation of shapefile at 1000m

df <- geolink_ntl(time_unit = "month",
                 start_date = "2020-01-01",
                 end_date = "2020-03-01",
                 shp_dt = shp_dt[shp_dt$ADM1_PCODE == "NG001",],
                 indicator = "avg_rade9h",
                 grid_size = 1000,
                 extract_fun = "mean")

#estimate annual night time luminosity for each household within a 100 meters
#of it's location

df <- geolink_ntl(time_unit = "annual",
                  start_date = "2020-01-01",
                  end_date = "2020-12-01",
                  shp_dt = shp_dt[shp_dt$ADM1_PCODE == "NG001",],
                  indicator = "average_masked",
                  grid_size = 1000,
                  extract_fun = "mean")
}
}
