% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_opencellid}
\alias{geolink_opencellid}
\title{Download OpenCellID data}
\usage{
geolink_opencellid(
  cell_tower_file,
  shp_dt = NULL,
  shp_fn = NULL,
  survey_dt = NULL,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  buffer_size = NULL,
  survey_crs = 4326,
  extract_fun = "mean",
  grid_size = NULL
)
}
\arguments{
\item{cell_tower_file}{A csv.gz file path downloaded from OpencellID.}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons}

\item{shp_fn}{A character, file path for the shapefile (.shp) to be read (for STATA users only)}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey i.e. a household survey with latitude and longitude values.}

\item{survey_fn}{A character, file path for geocoded survey (.dta format) (for STATA users only & if use_survey is TRUE)}

\item{survey_lat}{A character, latitude variable from survey (for STATA users only & if use_survey is TRUE)}

\item{survey_lon}{A character, longitude variable from survey (for STATA users only & if use survey is TRUE)}

\item{buffer_size}{A numeric, the size of the buffer for \code{survey_dt} or \code{survey_fn} in meters.}

\item{survey_crs}{A numeric, the default is 4326}
}
\value{
A processed data frame or object based on the input parameters and downloaded data.
}
\description{
Download OpenCellID data
}
\details{
This function processes downloaded OpenCellID data,
which provides information about cell towers and their coverage areas.
The return dataframe gives a count of cell towers within the shapefile area.
For the function to run, the user will need to set up an account on www.opencellid.org and download the
required csv.gz file for the country for which they wish to run the analysis. The file location of this
dataset (csv.gz file) should then be passed into the cell_tower_file parameter. Please see example usage below/
}
\examples{

\dontrun{
\donttest{

# Example usage
opencellid_shp <- geolink_opencellid(cell_tower_file = "C:/username/Downloads/621.csv.gz",
                                     shp_dt = shp_dt[shp_dt$ADM1_EN == "Abia",])


opencellid_shp_fn <- geolink_opencellid(cell_tower_file = "C:/username/Downloads/621.csv.gz",
                                        shp_fn = "tests/testthat/testdata/shp_dt.shp")


opencellid_survey <- geolink_opencellid(cell_tower_file = "C:/username/Downloads/621.csv.gz",
                                        buffer_size = 2000,
                                        survey_dt = hhgeo_dt[hhgeo_dt$ADM1_EN == "Abia",])


opencellid_survey_fn <- geolink_opencellid(cell_tower_file = "C:/username/Downloads/621.csv.gz",
                                           buffer_size = 1000,
                                           survey_fn = "tests/testthat/testdata/xy_hhgeo_dt.dta",
                                           survey_lon = "x",
                                           survey_lat = "y")
}}

}
