% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_mach.R
\name{geolink_pollution}
\alias{geolink_pollution}
\title{This function downloads monthly pollution data from Sentinel for specific areas and time periods.}
\usage{
geolink_pollution(
  time_unit = "monthly",
  start_date,
  end_date,
  indicator,
  shp_dt,
  shp_fn = NULL,
  grid_size = NULL,
  survey_dt,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  buffer_size = NULL,
  extract_fun = "mean",
  survey_crs = 4326
)
}
\arguments{
\item{time_unit}{A character, the time unit for the data. Default is "monthly". For now, no other time unit is supported.}

\item{start_date}{An object of class date, this indicates the start time for which the indicator will be pulled.}

\item{end_date}{An object of class date, this indicates the end time for which the indicator will be pulled.}

\item{indicator}{A character, the indicator of interest. Indicator must be one of aer-ai, ch4, co, hcho, no2, o3, or so2.}

\item{shp_dt}{An object of class 'sf', 'data.frame' which contains polygons or multipolygons representing the study area.}

\item{shp_fn}{A character, file path for the shapefile (.shp) to be read (for STATA users only).}

\item{grid_size}{A numeric, the grid size to be used in meters for pulling the vegetation data. Defaults to NULL.}

\item{survey_dt}{An object of class "sf", "data.frame", a geocoded household survey with latitude and longitude values (optional).}

\item{survey_fn}{A character, file path for geocoded survey (.dta format) (for STATA users only & if use_survey is TRUE) (optional).}

\item{survey_lat}{A character, latitude variable from survey (for STATA users only & if use_survey is TRUE) (optional).}

\item{survey_lon}{A character, longitude variable from survey (for STATA users only & if use survey is TRUE) (optional).}

\item{buffer_size}{A numeric, the buffer size to be used around each point in the survey data, in meters (optional).}

\item{extract_fun}{A character, a function to be applied in extraction of raster into the shapefile.
Default is "mean". Other options are "sum", "min", "max", "sd", "skew" and "rms" (optional).}

\item{survey_crs}{An integer, the Coordinate Reference System (CRS) for the survey data. Default is 4326 (WGS84) (optional).}
}
\value{
A processed data frame or object based on the input parameters and downloaded data.
}
\description{
This function downloads monthly pollution data from Sentinel for specific areas and time periods.
}
\examples{
\donttest{

# Example usage
df <- geolink_terraclimate(var ="tmin",
                           year = 2017,
                           shp_dt = shp_dt[shp_dt$ADM1_EN == "Abia",])
}
}
