% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_ntl.R
\name{get_annual_ntl}
\alias{get_annual_ntl}
\title{Download Annual Night Time Lights Data}
\usage{
get_annual_ntl(
  year,
  version,
  link_base = "https://eogdata.mines.edu/nighttime_light/annual/",
  indicator = c("average", "average_masked", "cf_cvg", "cvg", "lit_mask", "maximum",
    "median", "median_masked", "minimum")
)
}
\arguments{
\item{year}{An integer, for the year of interest (yyyy), one year at a time}

\item{version}{A string character, for now only "v10" is supported more to come later}

\item{link_base}{character, the link to the base website or database}

\item{indicator}{A character, specifying the specific indicator of interest. Options are
"average", "average_masked", "cf_cvg", "cvg", "lit_mask", "maximum", "median",
"median_masked" and "minimum"}
}
\description{
Functions to download night time lights data from NASA's Earth Observation
Group's (EOG) database. The downloaded rasters are collected for different
time intervals, daily, monthly and annually.
}
