% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_chirps.R
\name{get_month_chirps}
\alias{get_month_chirps}
\title{Download monthly rainfall chirp data}
\usage{
get_month_chirps(
  start_date,
  end_date,
  link_base = "https://data.chc.ucsb.edu/products/CHIRPS-2.0/"
)
}
\arguments{
\item{start_date}{An object of class date, must be specified like as.Date("yyyy-mm-dd")}

\item{end_date}{An object of class date, must be specified like as.Date("yyyy-mm-dd")}

\item{link_base}{the link to the daily chirps data (no need to change this)}

\item{resolution}{A character, gridded resolution of the data (options area "p05" i.e.
0.05 degree resolution and p25 for 0.25)}

\item{dsn}{data source name (a local folder where the data is to be stored)}

\item{cores}{the number of PC cores to employ in pulling the data in parallel}
}
\description{
Pull rainfall data from the CHIRPS data at monthly intervals for a specified period
The data is downloaded in raster format (.tif)
}
