% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_ntl.R
\name{get_month_ntl}
\alias{get_month_ntl}
\title{Download Monthly Night Time Lights Data}
\usage{
get_month_ntl(
  username,
  password,
  start_date,
  end_date,
  version,
  no_tile = TRUE,
  slc_type = c("vcmcfg", "vcmslcfg"),
  indicator = c("avg_rade9h", "cf_cvg", "cvg", "avg_rade9h.masked"),
  link_base = "https://eogdata.mines.edu/nighttime_light",
  shp_dt
)
}
\arguments{
\item{username}{A character, username on NASA's Earth Observation Group database}

\item{password}{A character, password on NASA's Earth Observation Gruop database}

\item{version}{A string character, for now only "v10" is supported more to come later}

\item{slc_type}{A character, adjustment type made to the NTL data either "vcmcfg" or
"vcmslcfg"}

\item{indicator}{A character, specifying the specific indicator of interest. Options are
"avg_rade9h", "cf_cvg" or "cvg"}

\item{year}{An integer, for the year of interest (yyyy)}

\item{month}{An integer, for the month of interest}
}
\description{
Functions to download night time lights data from NASA's Earth Observation
Group's (EOG) database. The downloaded rasters are collected for different
time intervals, daily, monthly and annually.
}
