% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hhgeo_dt}
\alias{hhgeo_dt}
\title{Nigeria GHS 2018/19 Data}
\format{
A \code{data.table} and \code{data.frame} object with 5116 rows and 48 variables:
\describe{
\item{hhid}{Household ID}
\item{dist_road2}{Distance to the nearest road}
\item{dist_popcenter2}{Distance to the nearest population center}
\item{geometry}{POINT location of each household}
\item{...}{For a complete list of variables, see \url{https://microdata.worldbank.org/index.php/catalog/3557/related-materials}}
}
}
\source{
\url{https://microdata.worldbank.org/index.php/catalog/3557}
}
\usage{
hhgeo_dt
}
\description{
The full household geovars dataset from the Nigerian General Household Survey 2018/19.
}
\keyword{datasets}
