% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgsd_utils.R
\name{import_from_rgsd}
\alias{import_from_rgsd}
\title{Import spatial data from NINA's GeoSpatialData directory}
\usage{
import_from_rgsd(name, layer = NULL)
}
\arguments{
\item{name}{Character. Relative path from \verb{/data/R/GeoSpatialData}, or full path.}

\item{layer}{Character. Required if reading from a GPKG or GDB with multiple layers.}
}
\value{
An \code{sf} object or a \code{SpatRaster}, depending on input type.
}
\description{
Loads vector or raster data using appropriate methods from \code{sf}, \code{terra}, or \code{sfarrow}.
}
\examples{
# Load a GeoTIFF (raster)
elev <- import_from_rgsd("Elevation/Fenoscandia_DEM_10m/Original/dem_10m_fenoscandia.tif")

# Load a shapefile (vector)
huts <- import_from_rgsd("Buildings/Norway_FKB_Buildings/Processed/N50_TouristCabins/Norway_TouristCabins_ETRS89_UTM_zone_33N.shp")

# Load from a full absolute path
# alt <- import_from_rgsd("/data/R/GeoSpatialData/Natur/N50/N50_NatureTypes.shp")

# NOTE: GeoPackage and GeoParquet files are not yet consistently supported in this directory
}
