% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_ntl_internals.R
\name{ntl_downloader}
\alias{ntl_downloader}
\title{An deprecated function used to download NTL rasters from the EOG website
This function was useful when EOG needed a username and password. Therefore
this is not really in use anymore unless the opportunity necessitates}
\usage{
ntl_downloader(
  client_id,
  client_secret,
  username,
  password,
  grant_type,
  token_url,
  data_url,
  shp_dt
)
}
\description{
An deprecated function used to download NTL rasters from the EOG website
This function was useful when EOG needed a username and password. Therefore
this is not really in use anymore unless the opportunity necessitates
}
