% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonalstats_plugin.R
\name{postdownload_processor}
\alias{postdownload_processor}
\title{A function to process downloaded raster into the final output in shapefile/geocoded survey}
\usage{
postdownload_processor(
  raster_objs,
  survey_dt,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  extract_fun = "mean",
  buffer_size = NULL,
  survey_crs = 4326,
  name_set,
  shp_dt,
  shp_fn = NULL,
  grid_size = 1000
)
}
\arguments{
\item{raster_objs}{a list of raster objects to be extracted from}

\item{survey_dt}{a sf/data.frame object, the geocoded survey}

\item{survey_fn}{a character, the filename of the survey (.dta file) (STATA users only)}

\item{survey_lat}{a character, for latitude variable name in survey_fn
(STATA users only)}

\item{survey_lon}{a character, for longitude variable name in survey_fn
(STATA users only)}

\item{extract_fun}{function to be extracted}

\item{buffer_size}{a numeric, radius of buffer for survey_dt (or survey_fn) units.}

\item{name_set}{a nameset which will be created for the variables}

\item{shp_dt}{an sf/dataframe with polygons}

\item{shp_fn}{a character, the name of shapefile for STATA users only (.shp/.gpkg file)}

\item{grid_size}{a numeric, size of the size of the grid in meters}
}
\description{
A function to process downloaded raster into the final output in shapefile/geocoded survey
}
