% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadplugin_landcover.R, R/geolink_master.R
\name{process_raster}
\alias{process_raster}
\title{Process a single raster (project to 4326)}
\usage{
process_raster(
  shp_dt = NULL,
  survey_dt = NULL,
  raster_file,
  extract_fun,
  survey_crs = 4326,
  name_set = "geolink_feature",
  weight_raster = NULL
)

process_raster(
  shp_dt = NULL,
  survey_dt = NULL,
  raster_file,
  extract_fun,
  survey_crs = 4326,
  name_set = "geolink_feature",
  weight_raster = NULL
)
}
\arguments{
\item{shp_dt}{A sf/dataframe with polygons}

\item{survey_dt}{A sf/data.frame object, the geocoded survey}

\item{raster_file}{File location of local raster to be used}

\item{extract_fun}{function to be extracted}

\item{survey_crs}{CRS for survey data, defaults to 4326}

\item{name_set}{A nameset that will be created for the variables}

\item{weight_raster}{a raster object used as weights when applying extract_fun, defaults to NULL}

\item{path}{Path to input raster}

\item{output_path}{Path to save processed raster}

\item{...}{Additional parameters passed directly to the postdownload_processor function}
}
\value{
Path to processed raster or NULL if processing failed
}
\description{
Process a single raster (project to 4326)

Function to extract local raster
}
\examples{
\donttest{

 #example usage
}
}
