% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_master.R
\name{run_geolink}
\alias{run_geolink}
\title{Master function to run geolink_chirps, geolink_ntl, and geolink_landcover}
\usage{
run_geolink(
  func_name,
  time_unit,
  start_date,
  end_date,
  shp_dt,
  shp_fn = NULL,
  grid_size = 1000,
  survey_dt,
  survey_fn = NULL,
  survey_lat = NULL,
  survey_lon = NULL,
  extract_fun = "mean",
  buffer_size = NULL,
  survey_crs = 4326,
  annual_version = "v21",
  month_version = "v10",
  indicator = NULL,
  slc_type = "vcmslcfg",
  use_survey = TRUE,
  buffer_survey = FALSE,
  iso_code,
  UN_adjst = NULL,
  constrained = NULL,
  bespoke = NULL,
  version = NULL,
  osm_feature_category,
  osm_feature_subcategory,
  buffer = NULL,
  stata = FALSE,
  var,
  res,
  model,
  ssp,
  time,
  key,
  year
)
}
\arguments{
\item{func_name}{A character string indicating the function to run, either "geolink_chirps", "geolink_ntl", or "geolink_landcover".}

\item{time_unit}{A character string indicating the aggregation period. For `geolink_landcover`, it must be "annual".}

\item{start_date}{A Date object or a character string in the format "yyyy-mm-dd" specifying the start date.}

\item{end_date}{A Date object or a character string in the format "yyyy-mm-dd" specifying the end date.}

\item{shp_dt}{An 'sf' or 'data.frame' object containing polygons or multipolygons.}

\item{shp_fn}{A character string specifying the file path for the shapefile (.shp) (optional).}

\item{grid_size}{A numeric value indicating the grid size to be used in meters. Defaults to NULL.}

\item{survey_dt}{An 'sf' or 'data.frame' object representing a geocoded household survey.}

\item{survey_fn}{A character string specifying the file path for the geocoded survey (.dta format) (optional).}

\item{survey_lat}{A character string specifying the latitude variable from the survey (optional).}

\item{survey_lon}{A character string specifying the longitude variable from the survey (optional).}

\item{extract_fun}{A character string specifying the function to be applied in extraction of raster into the shapefile.
Default is "mean". Other options are "sum", "min", "max", "sd", "skew" and "rms".}

\item{buffer_size}{A numeric value specifying the buffer size in meters (optional).}

\item{survey_crs}{An integer specifying the CRS for the survey data. Default is 4326.}

\item{annual_version}{A character string specifying the version of annual NTL data for download. Default is "v21" (for `geolink_ntl` only).}

\item{month_version}{A character string specifying the version of monthly NTL data to use. Default is "v10" (for `geolink_ntl` only).}

\item{indicator}{A character string specifying the NTL indicator to use (for `geolink_ntl` only).}

\item{slc_type}{A character string specifying the SLC type (for `geolink_ntl` only).}

\item{use_survey}{A logical indicating whether to use survey data (for `geolink_ntl` only). Default is TRUE.}

\item{buffer_survey}{A logical specifying whether to estimate a statistic based on distance from the survey location (for `geolink_landcover` only).}

\item{iso_code}{A character, specifying the iso code for country to download data from. (For 'geolink_population' and 'geolink_buidings' only)}

\item{UN_adjst}{A character 'Y' if want 2020 UN adjusted dataset returned, if not leave blank.(For 'geolink_population' only)}

\item{constrained}{A character 'Y' if want constrained dataset returned, if not leave blank.(For 'geolink_population' only)}

\item{bespoke}{A character 'Y' if want bespoke dataset returned, if not leave blank. Version also should be specified.(For 'geolink_population' only)}

\item{version}{A character such as v2.0 or v2.1 which correlates to the iso_code provided for bespoke datasets.(For 'geolink_population' and 'geolink_buidings' only)}

\item{osm_feature_category}{A character, refering to the osm key wiki page, please see details below. (For 'geolink_get_POI' only)}

\item{osm_feature_subcategory}{A character, refering to the osm key wiki page, please see details below. (For 'geolink_get_POI' only)}

\item{buffer}{buffer area around shapefile. (For 'geolink_get_POI' only)}

\item{stata}{A flag for stata users. (For 'geolink_get_POI' only)}

\item{var}{A character, the variable of interest (e.g., "temperature", "precipitation").(For geolink_CMIP6', geolink_worldclim' and 'geolink_terraclimate' only)}

\item{res}{A character, the resolution of the data (e.g., "2.5m", "5m").(For geolink_CMIP6' and 'geolink_worldclim' only)}

\item{model}{A character, the climate model name (e.g., "ACCESS-ESM1-5", "CanESM5").(For geolink_CMIP6' only)}

\item{ssp}{A character, the Shared Socioeconomic Pathway (SSP) scenario (e.g., "ssp126", "ssp585").(For geolink_CMIP6' only)}

\item{time}{A character, the time period of interest (e.g., "historical", "2020-2049").(For geolink_CMIP6' only)}

\item{key}{A character, the API key created in when signing up to Opencellid profile. (For 'geolink_opencellid' only)}

\item{year}{A numeric, the year for which data is to be downloaded. (For 'geolink_terraclimate only)}
}
\description{
This function serves as a master function to run the functions in geolink_mach, each function based on the provided function name.
It dynamically passes all other arguments to the selected function.
}
\examples{
\donttest{

# Example usage
   df = run_geolink(
     func_name = "geolink_chirps",
     time_unit = "month",
     start_date = "2020-01-01",
     end_date = "2020-03-01",
     shp_dt = shp_dt[shp_dt$ADM1_PCODE == "NG001", ],
     grid_size = NULL,
     survey_dt = hhgeo_dt,
     extract_fun = "mean")


     df = run_geolink(
     func_name = "geolink_electaccess",
     shp_dt = shp_dt[shp_dt$ADM1_EN == "Abia",])


}




}
