% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolink_master.R
\name{run_geolink}
\alias{run_geolink}
\title{Master Function for Geospatial Data Extraction}
\usage{
run_geolink(
  data_type,
  start_date = NULL,
  end_date = NULL,
  iso_code = NULL,
  shp_dt = NULL,
  survey_dt = NULL,
  ...
)
}
\arguments{
\item{data_type}{Type of data to extract}

\item{start_date}{Optional start date}

\item{end_date}{Optional end date}

\item{iso_code}{Optional ISO country code}

\item{shp_dt}{Optional spatial dataframe}

\item{survey_dt}{Optional survey points}

\item{...}{Additional parameters passed directly to the specific data extraction function}
}
\description{
Master Function for Geospatial Data Extraction
}
