% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shp_dt}
\alias{shp_dt}
\title{Boundary Shapefile for Nigeria}
\format{

}
\source{
World Bank subnational boundary repository
}
\usage{
shp_dt
}
\description{
The full shapefile of Nigeria up to the second administrative divisions i.e. the local government
areas
}
\keyword{datasets}
