% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonalstats_plugin.R
\name{zonalstats_prepshp}
\alias{zonalstats_prepshp}
\title{A function to process raster downloads and compute zonal statistics for shapefiles
and geocoded surveys}
\usage{
zonalstats_prepshp(shp_dt, shp_fn = NULL, grid_size)
}
\arguments{
\item{shp_dt}{a sf/data.frame object, the shapefile}

\item{shp_fn}{a character, the name of shapefile for STATA users only (.shp/.gpkg file)}

\item{grid_size}{a numeric, size of the size of the grid in meters}
}
\description{
A function to process raster downloads and compute zonal statistics for shapefiles
and geocoded surveys
}
